[roll] Roll fuchsia [kernel][ktrace] Sidestep ktrace importer limitation.
Sidestep a limitation in the ktrace importer implementation where
it fails to associate a thread id with its name and process id if
the importer encounters a record referencing the thread id before
encountering the name/process records for that thread. This is
caused by the importer adding a default record to its thread map
whenever it encounters an unknown thread id. The map is not updated
later when the name/process records are encountered.

This issue manifests in Chromium Trace Viewer as numbered swimlanes
showing up in the "kernel" process, as the default process id the
importer uses is 0 when encountering unknown thread ids.

This issue could be addressed in the importer by either:
1. Updating the thread map whenever the name/process records are
   encountered.
2. Changing the importer to process the event buffer for name/
   process records before other events.
3. Changing the kernel to emit the live thread/process records
   before enabling tracing.

This patch takes option #3, as it is the simplest and has no
artifacts. Option #1 would still produce swimlanes in the kernel
process for as many records are emitted before the name/process
records are emitted. Option #2 is much more involved.

Test: Take 10 traces and observe no artifacts in Trace Viewer.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480359
Original-Revision: bcdcdf4e4c94d958031956b6f218acd0d4ee9fca
GitOrigin-RevId: a4e387f3dbf27a07b2062783ccb48f6054b2f3d5
Change-Id: I29d38751553023a1cc73aee709585ae4aafdebc3