[roll] Roll fuchsia [fshost] fix diagnostic directory forwarding
fshost hosts a forwarded version of the blobfs diagnostic directory in
it's own diagnostics directory, but it was broken with the change to
using export directories to route those handles for the filesystems.
This fixes the forwarding for diagnostic directories.

Test: fx test fshost-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/479305
Original-Revision: 911b4f8537a1b8f63b70d7b29f1a88f612acc544
GitOrigin-RevId: b85aef3945fbc11b0cdc1e2d0f446675235ce8e8
Change-Id: I8ad106b3e5258d25dac359601d6932d19b4221a3