[roll] Roll fuchsia [audio_core] Fix overflow resulting from a rounding error
For justification for this change, see this comment:
https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=67996#c22

This CL adds a regression test which fails before this CL (by crashing
in PointSampler) and passes after this CL. The test I added relies on
making a direct call to MixStage::ProcessMix after directly setting some
values Mixer::source_info and Mixer::bookkeeping. This is ugly -- it
would be nicer to use the MixStage public API only -- however MixStage
has so much internal state that it's difficult to create a test case
which makes that internal state align in exactly the right way to
trigger the crash.

Original-Bug: 67996
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474704
Original-Revision: cd67dcc480c733d6ec5ba1be525774ceac232a89
GitOrigin-RevId: 71768a04aaaa8b8e553156f31872f671c036820b
Change-Id: Ida0f29af2962258b13c7cead44781fd0d0786a05