[roll] Roll fuchsia [perfcompare] Specify precision explicitly for printing point estimates
This preserves the behaviour that Python 2 gives and hence keeps the
golden file tests passing when they are run on Python 3 (with other
fixes applied).

On Python 2.7, str() on a float is equivalent to formatting with
'%.12g' while repr() on a float is equivalent to formatting with
'%.17g'.

By Python 3.8, that has changed so that str() and repr() on a float
both use '%.17g'.

Python's behaviour was changed in this issue:
https://bugs.python.org/issue9337

The older behaviour was documented here:
  "repr(float) produces 17 significant digits"
  "Python’s builtin str() function produces only 12 significant digits"
https://docs.python.org/3.0/tutorial/floatingpoint.html

Original-Bug: none
Test: ./prebuilt/third_party/vpython/vpython garnet/bin/perfcompare/perfcompare_test.py
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481023
Original-Revision: 2ad8f8e99de88412113e1b33691c9431b7b50749
GitOrigin-RevId: 8550ad9c0ba41313fa7041f362ee355811160b48
Change-Id: I247e3adf1d69d8b7048d3e23238f53c9da080344