[roll] Roll fuchsia [build][tracer] Failures should not freshen output
Check that failed actions do not leave an output that appears
to be up-to-date relative to inputs.

Note that depfiles at this time are still allowed but not required,
which means that writes to depfiles from failed actions will not be
caught with the current logic.

For reference, automake does this safely by writing a depfile to a
temporary file and then moving it in-place after the compile command
succeeds.  This way the depfile always reflects the last good compile.

This is not expected to have any impact on healthy/green builders,
but may alert users with locally failed actions that something is wrong.

Test: fx test build/tracer

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477605
Original-Revision: 625bfd5c0de20100a8c60a598007d72e949e51ef
GitOrigin-RevId: 6f556d1417f31c384894ea5e59352d192454f813
Change-Id: Ida0256f3e96d5c0818f7b7b3a42340b63da82902