[roll] Roll fuchsia [stdcompat] add cpp20::bit_cast.
The bit_cast polyfill is a 3-way branch, were std::* is picked first,
if this is not available we use bit_cast builtin. As a last resort we
introduce a non constexpr bit_cast.

The usage of aligned_storage, for memcpy and then reinterpret_cast,
allows us not to introduce a default_constructible requirement on type
To.

Introduced a LIB_STDCOMPAT_NONCONSTEXPR_BIT_CAST, defined only when the
last resot branch is picked. This is mostly for testing constexpr
scenario when this is not defined.

Original-Bug: 57356
Test: stdcompat-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478620
Original-Revision: 2154cf4a4493d4de9b42efc86b63dfebffb7168c
GitOrigin-RevId: 51c5e878c6dbf32e68ed9f2c44ea269844580c06
Change-Id: I82f567fdacbfe4499bc59cc572abdb63bf1ae0f3