[roll] Roll infra/recipes [testing] Delete extra logs from failed test steps
The recipe engine adds "stdout [raw]" and "execution details" logs to
every step that runs a subprocess. This applies to the steps that
present failed test names and logs because because we use that step to
read the log from disk.

However, it's confusing to have those logs (especially "stdout [raw]")
for a step that's supposed to represent a failed test that ran on
another machine. So instead, read the test stdio files proactively while
processing results and then present them in non-subprocess steps so that
the steps don't have "execution details" or "stdout" logs.

Original-Bug: 68488
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/481592
Original-Revision: cbfbf459c34f541cbee6aaffd1f5e7578ee0c9d8
GitOrigin-RevId: 9cab6cddf6f177e1a5aafd3054a11618c2af8678
Change-Id: I1f71ec8b36db27d715fd8df465431d32bbe3bc10