[roll] Roll third_party/go Rely on files rather than directories
This test compares the size reported by `stat` to the number of bytes
that can be read by `open`ing the path and `read`ing from it. As
`/pkg/{bin,lib,meta}` are directories, and evidently `stat`ing
directories disagrees with the number of bytes that can be read from
them, this is not suitable test fixture on Fuchsia.

Use the test binary itself to avoid this problem.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/482237
Original-Revision: 53ac1bf8e885237d5203183f31ff1156b2ea04c0
GitOrigin-RevId: 72d990fbb89072dfc4b9da2bf6081251cfa9f42e
Change-Id: Ia1e1b18eba4e189b45ce2283de3cf65c5f307d0d