[roll] Roll fuchsia [uuid] Add ostream output operator.
This is very useful for debugging. operator<< for streams is now allowed
by the style guide for "value" types (which Uuid is) and this cleans up
the call sites slightly for code that uses streams for output.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/479222
Original-Revision: 930c8ba7f78a1c07df792e73414d5a5d50c80ec6
GitOrigin-RevId: 5ad4fc924ba57b383291d72f6e74aa14e794b796
Change-Id: I00ca9377f9cf59ce47a721d897f1a191c2674510