[roll] Roll fuchsia [bind] Handle match in new bytecode
Made changes to bindc so that it'll only add a
UnconditionalBind instruction to the old bytecode.

Made changes to encode_bind_program_v2 so that
it'll return an error if it encounters a Match
instruction.

Original-Bug: 67441
Tests: Added a unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481579
Original-Revision: 84af962f255b2f3731f954e7fde7f9aa72554182
GitOrigin-RevId: 99b2e6353051ddbd5f5a73cc5af581235b402b69
Change-Id: I277453d035499636766ac8bac8e40808fcea2055