[roll] Roll infra/recipes [cherry_pick] Allow atomic, bulk cherry-picks
Add a new property `changes` which is the bulk version of `revision`
and `change_id`. Run cherry-picks in a loop until completion, or a merge
conflict is hit. In case of a merge conflict, abort.

All changes in the stack are required to share the same project, e.g.
only allow cherry-picks of all fuchsia.git changes, but not of a
fuchsia.git change + a topaz.git change.

Original-Bug: 69241
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/482322
Original-Revision: a6d67a594aee0797326a6ba538736c45544fcfe5
GitOrigin-RevId: dabd729042edb34e0bf570c81e79ff497ef154fb
Change-Id: I3c34f86c9d6c84cf61c5fdb79f1f09157fd201a0