[roll] Roll fuchsia [fshost] Proxy BlobfsVerifier FIDL from blobfs
Exposes the fuchsia.update.verify.BlobfsVerifier FIDL interface out of
fshost, which is implemented as a proxy service node.

Refactored fs-manager to not take ownership of the root export channel
in ForwardFsDiagnosticsDir, so that we can store the root export channel
in FsManager and reuse it for as many service proxies as we'd like.

Original-Bug: 64608
Tested: Integration test
Tested: Manually verified the service is exported: `cs info -f fshost`
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481625
Original-Revision: fc8991118e1e690868eaccd4f59acd3918fd7181
GitOrigin-RevId: 20a5f58f14aac280205f14110c45a9c896ad7f33
Change-Id: Ib628c734e5a94e2458e9d9a5121f58ef3d49ce3d