[roll] Roll fuchsia [SetUI] Use Rebroadcast in Policy.
Currently the policy handler generates its own
settings change event in order to indicate when the
value has changed due to changes in policy. This is
coupled tightly with the switchboard logic, which is
being deprecated.

This change moves the notification responsibility to
the setting proxy via the rebroadcast command. The
setting proxy is then able to use its standard
communication method to convey the change.

The service MessageHub has been plumbed to the
policy handler as part of this change so that the
policy handler can invoke the rebroadcast command
on the proper MessageHub.

Test: fx test -o setui_service_tests
Original-Bug: 68489
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481011
Original-Revision: 555ad20a781365b4beecfb7d42cd9a047ff45ec0
GitOrigin-RevId: ff297bb18addc6aeed43cd0825170152ce50ba5f
Change-Id: I68d61a4b9d645cf13da160be352fda78fc74c39a