[roll] Roll fuchsia [go] Commit go.sum
It's essentially noise given our configuration but it's required by go1.16.
To clarify, we map Go dependencies via jiri (many are forked) and use
replace statements in go.mod.

I confirmed the file to be up to date with:
  for i in */go.mod; do G=$(dirname $i); echo $G; cd ./$G; go mod tidy; cd -; done

and confirmed resolving the go1.16 induced error with:
  cd tools
  # Edit go.sum and remove all references to cloud.google.com/go/storage
  go mod download cloud.google.com/go/storage

and the file became as before.

For more information see
- https://github.com/golang/go/issues/41934
- https://github.com/golang/go/issues/40728

The TL;DR: is that -mod=readonly becomes the default and when it's set,
go.sum must be present and clean.

This change has no effect on the GN build.

Original-Bug: 69326
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482481
Original-Revision: 0cfd5baeac6d74287c2c277ef0efd532b82fce64
GitOrigin-RevId: 648ee61f3a4a67802c6b630edbbd651e9a395187
Change-Id: If65d2e576b9c23eb2c4211e8534f5d022b0f848d