[roll] Roll fuchsia [kernel][arm][gic3] fix an out of range bug in the SGI delivery
This only happens if a SGI is attempted to be delivered prior to
a secondary coming online. The previous logic would end up looping
beyond the end of a cluster id and could exceed the range of
SMP_MAX_CPUS when calling arch_cpu_num_to_cpu_id();

Add a terminating condition to the inner loop and some debug asserts
in the accessor routines, as well as rearrange some code and add some
trace printfs.

NOTE: this is somewhat of a bandaid and a somewhat cleaner solution
can be made with the way we track cluster (AFF1) and cpu ids (AFF0)
within the cluster to avoid the inner loop.

Original-Bug: 68976
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480361
Original-Revision: c2177b1653b522b0e408fedfbbac19978e470975
GitOrigin-RevId: cfe010e43e93801a2d29e618fe5499f3c79165c2
Change-Id: I7c1ee8fcf9a979dfae4676e4bb11de87a04e121f