[roll] Roll fuchsia [kernel][phys] Refactor memory allocation
Use a bespoke owning type rather than unique_ptr specialization.
The custom deleter object storing the size means the UniquePtr
object is already two words and knows the size, but not in a
convenient fashion.  This made a separate object holding the size
be three words instead of two, with the same size stored twice.
The bespoke object just holds it as a span and makes it easy to
access so no separate object is needed, while also handling the
ownership semantics.

The Allocation object is now the only entry to the allocator API.
It uses fbl::AllocChecker to enforce error checking.

This separates the general allocation code from the code specific
to reading the ZBI, so it can be used in different phys contexts.

Original-Bug: 68762
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478857
Original-Revision: 33f0505d84569475ffcd7726176c191db71697d9
GitOrigin-RevId: c145476ad3ee2304e044f4f6ecdc0e9eed2dd521
Change-Id: Ib811b81c13bb369fdb3bf7b81c4bd1e30d8779ff