[roll] Roll fuchsia [kernel][phys] Decompressing ZBI chain-loader
This adds tests of decompressing a whole ZBI embedded in a ZBI item
and booting that ZBI.  So far it only works for arm64, where no page
tables are needed.  This is a more strenuous test of decompression
as physboot will use it, and it's also a test of arch::ZbiBoot that
boot shims (and interim physboot) will use.

Original-Bug: 57764, 68762
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478720
Original-Revision: 05897844e2518a99facafc538006ef2993f64226
GitOrigin-RevId: 37f13c64c1ab01c805121f771c63ef921e820aa5
Change-Id: I003889b83eeba494a49bd2e898e532c5cb5741bf