[roll] Roll fuchsia [media][audio_core] Update fuzzer deps
The audio-fuzzer/api-fuzzer is crashing due to missing deps and services
that bring in the fuchsia.media.AudioTestDevmgr service.

Locally, these failures are masked due to the inclusion of the services
and deps in encompassing packages; because fuzzer packages are extracted
as isolated units, these services and deps need to be explicitly
included.

Original-Bug: 65547

Test: fx fuzz audio-fuzzer/api-fuzzer
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482657
Original-Revision: bb5c0843cf107068904e1689a0adb0965ba0464b
GitOrigin-RevId: 8026025928525ed3a8adb306bd0d7fb0f2af7206
Change-Id: Ie4b08c65df18661cc162b46e61d5bddf8be7e98e