[roll] Roll fuchsia [nelson][selina] Add selina device in its own devhost
This prevents an issue where the selina driver could become stuck due to
it being in the same devhost as the GPIO driver. SPI uses GPIO as well,
so there is a circular dependency if the selina and GPIO drivers are in
the same devhost.

Original-Bug: 67567
Test: Radar CLI no longer hangs waiting for frames
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482497
Original-Revision: 674c5e761d92b2ff268a42cad89b8fa4944b4f0d
GitOrigin-RevId: 8c294cf9e49e51e843759ca66a6270def567c328
Change-Id: I9cb05599a465857e1c818c156afd06b0d79656e7