[roll] Roll fuchsia [debugger] Syntax highlight source code listings.
Uses the tokenizer to provide simple syntax highlighting in source code
listings. This helps improve code readability significantly.

The syntax highlighter is limited to what the current tokenizer
supports, so there are no comments or C character literals, but this
will improve as we enhance the tokenizer.

A list of keywords for Rust and C is added since the tokenizer doesn't
split apart each keyword as a separate token.

Fixes a bug in the tokenizer where an underscore was treated as a
separator if immediately following a keyword. A test is added for this.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482682
Original-Revision: d18bdc59157f5f7bb68a9bfa3b5a9dede20852ea
GitOrigin-RevId: 5d2053237dbc9577c7f6c0054a4715a120a8d09a
Change-Id: I519f6b4e488d9dccde1be7360b7cd530503cf91c