[roll] Roll fuchsia [fidl][cpp] Generator name handling rework
This is a significant set of changes to the way that C++ type names,
declaration names and namespaces are handled in the C++ fidlgen
backends. It's a single CL because the changes are fairly tightly
connected. There will be follow-up CLs that build upon this one to
transition more name handling to the patterns developed here.

This is motivated the C++ binding unification project. This will require
code generators to refer to both the wire (aka LLCPP) and natural (aka
HLCPP) types in the same files. This means that we need an IR that knows
the different names for types of generated type.

GO IR CHANGES:

Namespace is a string slice that represents a C++ namespace. The
compiler has a set of functions to turn a FIDL library name into the
appropriate namespace for a particular kind of type.

DeclVariant holds a C++ declaration name: a namespace (as a string
slice) and the name within that namespace. DeclName holds two
DeclVariants one for the Wire variant, one for the Natural variant.
There are methods on DeclVariants and DeclNames to manipulate the
namespace and the name.

The compiler has a compileDeclName method that returns the base decl
names (wire and natural) for a given FIDL declaration.

TypeVariant is a string that represents a C++ type. TypeName holds two
TypeVariants, for Natural and Wire types. There are methods on TypeName
and TypeVariant to manipulate the types, primarilly to wrap them in
C++ template invocations.

TEMPLATE CHANGES:

Wherever possible when referring to a type or a value we now use the
fully qualified type or declaration name, starting at the global
namespace.

This can't be done everywhere, for example a class must be declared in
its actual namespace so there's a utilty function {{ EnsureNamespace }}
that takes the namespace (or DeclVariant) that the next code should be
generated at and ensures that the generated C++ code will be in that
scope. To do this fidlgen tracks what namespace it thinks it should be
in.

ifdefs make this a bit tricky so there are {{ PushNamespace }} and
{{ PopNamespace }} helpers to ensure that the namespace state is
consistent when the code is compiled for host and device. There's also a
{{ EndOfFile }} helper that verifies that the namespace stack is empty
and then closes any open namespaces.

Original-Bug: 65964

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480962
Original-Revision: 5c791fde10e60b408bc08325e39e6559dd197e0d
GitOrigin-RevId: 8f3db83a1ba73815ab695cded89f39fc476800d4
Change-Id: I37b4ebde228bec99dc30cfb7ef5f8a551a1d3c71