[roll] Roll fuchsia [fidl] Generate dangerous ident tests at build time
This removes the checked-in generated dangerous identifier tests. All of
the tests are now generated at build-time.

Now that tests are checked in we can drop the 'tag' from identifier
definitions - they were just there to minimize diffs for the checked in
tests. Without the tags it's possible to shard the tests so that each
individual generated type is smaller - so this shards them into four.
With sharding it's possible to re-enable some tests that were previously
disabled because of build performance.

The deny listing is now almost entirely done in identifiers.py, but
there are some issues with Rust bindings deny listing and structs so the
struct tests are explicitly disabled for Rust in BUILD.gn

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443697
Original-Revision: f4e72e96deee8c960a63ce6f78cc1278845878d8
GitOrigin-RevId: 2018bdec06a6b6a62f769b1e935a7abb4b44d73c
Change-Id: I790077fdff03269292501e1d7af89d1209071629