[roll] Roll fuchsia [kernel][vm] Support discarding pages
Implement discard operation that frees up all the pages of a reclaimable
discardable VMO, and sets the state to "discarded". Future unlocked
accesses on the discarded pages will fail. This is implemented by
checking the current discardable state in GetPageLocked(), and failing
the call if the state is discarded. VMOs will transition from the
discarded state to the unreclaimable state on the next lock operation.

This commit just introduces the ability to discard a VMO, but that
functionality is not used yet outside of tests.

Test: kernel-unittests

Original-Bug: 60237
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476158
Original-Revision: cb1c6c41c38df63c6e8e4c5ac0d38b71db2d0281
GitOrigin-RevId: 6219fb955633fcf2d91c55c3cb08f091fb5d32fc
Change-Id: Id6ff01affe5af5e3af62e3d92ecd4bfa7e8a9318