[roll] Roll fuchsia [kernel][vm] Query discardable page counts
Query locked and unlocked page counts over all discardable VMOs in the
system. This requires iterating over the global discardable lists of
VmCowPages, which requires the DiscardableVmosLock; and querying the
page counts on each object, which requires the VmCowPages lock. Since we
cannot acquire the VmCowPages lock while holding the DiscardableVmosLock
(to avoid lock ordering deadlocks), we use the same strategy as for
traversing the all_vmos_ list.

The VmoCursor class is templated so it can be used both for all_vmos_
and the discardable lists. The DiscardableVmosLock is dropped between
iterations of the discardable lists, so that each individual VmCowPages
can be queried outside of the DiscardableVmosLock.

The discardable page counts can be printed with 'k scanner dump'. They
will also be exposed later through kmem stats.

Test: kernel-unittests

Original-Bug: 60237
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478497
Original-Revision: bb14387258a4a1652de89543c15ef47865b7f707
GitOrigin-RevId: b9f276376c713f6b80c936588baadd5bc848be51
Change-Id: Id050ee8536ca29f44c3b3c796f7b38739840e3ab