[roll] Roll fuchsia [lapis] Rename "EventCount" to "Occurrence".
EventCount in Cobalt v1.0 is an overloaded type that
acts as a bucket for the number of times a thing happens, and
for raw integers. In v1.1 this ambiguity is addressed by introducing
explicit "Integer" and "Occurence" types. This naming is much clearer
and Lapis should use it to make the V1.1 cobalt migration more apparent.

Original-Bug: 42067
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483240
Original-Revision: c45caeb1fdc6cfee712c711afec35a291c9ed2ea
GitOrigin-RevId: 88ccd5d742246fa3fa30936a091b7c1cca7de964
Change-Id: Iad165ae1c5aeae9aeabb439a3de019cc2579cf7a