[roll] Roll fuchsia [debugger] Fix some TLS printing problems.
TLS access was broken when ELF symbols were required to be explicitly
annotated as such to be found by symbol lookup. This annotates the TLS
symbols as being ELF symbols.

Fixes an asan use-after-free error around the DwarfExprEval that is used
to execute the TLS program. This is unfortunately delicate that the
DWARF TLS thing can't conveniently use the simpler ASyncDwarfExprEval
class that would avoid this problem.

This TLS integration can't easily be tested without an end-to-end
integration test which is not currently practical.

TLS is still broken on x64 due to the fs.base register being missing.

Original-Bug: 53940
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483017
Original-Revision: e5ba1c994b624a75fbdf07ba69140c18c836c0c8
GitOrigin-RevId: e868f995e801143c3574fe71091c6e451abde07d
Change-Id: I1bce606fec231a56f22c9591732993e7c06d9533