[roll] Roll fuchsia [journal] Change the fuzzer to avoid hang
The journal's current API means that if you call WriteData and there's
not enough room in the buffer to store the operations, it will block
until there is room, but if the caller hasn't done anything with the
previous reservation, it will just deadlock. To fix this properly, we'd
need to substantially modify the journal API so that it keeps track of
which reservations have been submitted and which haven't been, but
that's tricky.

For now, just fix the fuzzer so that it doesn't send two data
operations, only one, and that should never block.

Original-Fixed: 66376
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483011
Original-Revision: 81f090bc7e9b45eb9aad13a6ce37c695df0eb67c
GitOrigin-RevId: 8c011f53231ff8370963ab1d970136d847e9fddb
Change-Id: Ieb5a427aed1fe5447a3de865b591a3f4ae7c2afe