[roll] Roll fuchsia [scrutiny] Add configurable build path.
This change adds a configurable build path option to Scrutiny. This
allows the build path to be fixed instead of retrieved the environment
which is useful for build scripts or custom builds. Setting the
build_path will let you define where the model collects data from.

This change also integrates a new testing library into Scrutiny which
removes a lot of duplication of the testing code and was required to
simplify testing the new ModelEnvironment integration. In addition this
change cleans up package data collection to improve its readability
which was one of the core consumers of the build path from the
environment.

Original-Bug: 68938
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481016
Original-Revision: 782308531b05848f934a040745bf7ba66a641e94
GitOrigin-RevId: 9451d51c73eeda70883d975a624265475f158064
Change-Id: I58b068c19fcad68a88981ea21a36549515c685c0