[roll] Roll fuchsia [rustc_binary] make output_name support "-"
If output_name is passed to rustc_binary, that becomes the output_name
of the executable, without "-" -> "_" normalization, as one might expect.
The crate_name of the binary (observable with `module_path!()` )
continues to be normalized.

A future CL will apply this same logic to the target name itself, but
output_name is part of the transition path.

Original-Bug: 64100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/450757
Original-Revision: 08e001ca339711bf1bc65cf7b86c9682e7c03df7
GitOrigin-RevId: cbe7ccec5380b2db202473bbad5569d8298a6f9e
Change-Id: Iea629acfea81e3ee9fdf3dc60b8a5ef5ee975f36