[roll] Roll fuchsia [scrutiny] Add command engine.model.environment
This small change adds a controller that exposes the model environment.
This is useful for debugging consumer issues when they aren't getting
their expected results from their build as we can verify through a
command that the model environment is setup correctly. Such as the build
directory is at the correct path and the data models uri is correct.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482198
Original-Revision: 45b5de3ccd439e8daa960bce34174247c654f5d9
GitOrigin-RevId: 3d17579242616389043708da2312cf155a169628
Change-Id: I5c054591ec46e67bbf5a5fc4c275bce65ac7d1ad