[roll] Roll fuchsia [scrutiny] Add meta to core::Package collection definitions.
Currently packages only support contents which isn't all of the data
available in a package. Inline metadata can also be included directly
into the FAR. This data should be captured by the Package collection as it is
important for analysis. For example quickly examining all the allowlists
in the meta/ of a package.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482199
Original-Revision: e798e73dfc781092801da6fe6aa85199cccf27a2
GitOrigin-RevId: 9cbc4fa078f8ae27e4e1b31d56a2a7407abdfa58
Change-Id: Ie66dc791c7f51ca98c5cede55b08a7d3e5361975