[roll] Roll fuchsia [pci] Add kPCI bus mastering enable missing in 6e34cb3dbcc
The previous CL made it through before noticing the final change to
actually enable the support didn't make it into the commit. This adds
it and verification was done with the i915 changes to ensure it's
working.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483341
Original-Revision: 100288e3f9a76f6b7e5e82eb3f7685a0c835fa62
GitOrigin-RevId: 14957bfa1d7871a3c9b85101f90a8901dbfbec30
Change-Id: Id0d157f32315bf4e5bb567cac4e1856d122cbba9