[roll] Roll fuchsia [libc] Build workaround to unstick gwp_asan roller
The gwp_asan sources removed a file, so the next integration roll
can't land until the build is changed.  But the build can't be
changed to omit the old file until the inregration roll lands.
So split the difference with a kludge to detect the old pin state
and adapt the build to both old and new.  The conditional hack
can be removed as soon as the integration roll lands.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483349
Original-Revision: 0802472018e4701300bf56e25ccf9bc36fa9d397
GitOrigin-RevId: b6c30254cbb321a0ed098eac078fa74e36268f29
Change-Id: Id3bf455fda8c611c2ba082f5d3f2a57f180cf0f7