[roll] Roll fuchsia [stdcompat]: Add to_address polyfill.
std::span spec relies on to_address for multiple places, so I'm using
adding a polyfill for that. If its available we'll use that.

Noticed that we were not using the polyfills in on test, so it builds
even when the std version is not available.

Test: stdcompat-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482899
Original-Revision: 5ae9f3de0d0c596c0c4aff7eeadf2b45880884cf
GitOrigin-RevId: 92f57475436b86c72d5098ce284c38c286b4de88
Change-Id: Ifad948081c082584744ccd955106bfb7552ee224