[roll] Roll fuchsia [archivist] Switch to service config files instead of args.
We want to support different sets of diagnostics analysis services on
device depending on system build parameters. This change makes the
Archivist look in a specified location for a new services_config.json
file, which contains a list of services to connect to.

We connect to these services when Archivist starts up to support "lazy
binding." With this change, we will be able to create product
configurations that exclude certain services.

Original-Bug: 69296
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483460
Original-Revision: 5ecb25f5288431bd51bdc628d1634d3eb892b6c3
GitOrigin-RevId: 087553e85e5f561c4913aca2328735ab44d4c85b
Change-Id: I12b3ac027d3746a31b7d7b6a3919be7c6227f2b5