[roll] Roll fuchsia [fidl][fidlconv] Add alias converter to fidlconv.
The only thing that needs to be converted in an "alias" declaration is
the order of names/types, which is already accounted for by the existing
code.  This change mostly just adds new tests to validate that this is
proceeding as expected.

Test: fidldev test fidlc
Original-Bug: 68667
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481963
Original-Revision: 7a47f3e8c0fb5abefe1baecb447fbba2971643a8
GitOrigin-RevId: b579fc61826493db9a7d2e71d061ee0af0e9f007
Change-Id: Ifadf9e3e076ae598246843e361a09a319ebe2d7b