[roll] Roll fuchsia [cm][resolvers] Send fuchsia.mem.Data instead of ComponentDecl in Resolver protocol
Component manifests have lots of repeated string data, and can grow to
be larger than a zircon channe messagel size. Therefore, it is not a good
idea to send the ComponentDecl FIDL directly in a FIDL response.

Component resolvers typically deserialize the ComponentDecl from the
component's package, but don't actually read it. So it seems just forwarding
the uninterpreted bytes is the best solution. To account for large
manifest sizes, we use fuchsia.mem.Data, which is a union that can
either contain the bytes inline, or a handle to a VMO.

For base-resolver and universe-resolver, this works well, since the VMO
can be extracted straight from the package and forwarded to component
manager, not needing to read the contents in the resolver.

For future resolvers whose components don't have packages, and whose
manifests are small, the resolver can send the manifest inline.

Original-Fixed: 69312
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482757
Original-Revision: 66f9f9058c34b5ac79fe2246af434e014f828f91
GitOrigin-RevId: fadccd3b0c219f82102ce83db40a978e7c86edcb
Change-Id: I98ab4a9644e9a7fbd330cb210cee5c57b33299d0