[roll] Roll fuchsia [banjo][c] Separate helper declarations from other declarations.
The ordering of protocol declarations is looser in FIDL, meaning a
protocol A may be declared before a protocol B it depends on.
Since helper methods require full types to be available, move these
methods to the very end of the generated header so that all necessary
declarations are available.

Original-Bug: 68157
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483381
Original-Revision: 627868dd2b78011cfa316618f295237fefbe3acd
GitOrigin-RevId: ffbbb5c9786fd4d1b94b8b94aeb211372ce918ab
Change-Id: I6150b9a1a6c9df64c563cb9ef676e2bae148a129