[roll] Roll fuchsia [blobfs] Fix unlink behaviour while clones exist
Blob::Purgeable was missing a check that there are no active VMO clones.
Unlinking the blob when there are no active VMO clones is unsafe since
the pager is holding onto a reference to the node index for the blob
(which would be unallocated by the unlink).

Also made AllocatedExtentIterator defensively check if the inode is
allocated, which made testing easier (otherwise the failure would be
silent, as the old inode was not overwritten but simply marked
unallocated, so still had good data). This required a small change in
Blobfs::FreeInode, which was creating an allocator *after* freeing the
inode (although before persisting that change).

Added a new unit test which reliably triggers the issue.

Original-Bug: b/177017725
Multiply: blobfs-unittest
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483499
Original-Revision: a9c60c830cf1eae75c3c1d8ea7f3f640d449c170
GitOrigin-RevId: 824f6699330f3994e38a279d30d3772f3e6a97f7
Change-Id: I0b56350ae3a5bc106e19d0dd7b0ff464bd1890b5