[roll] Roll fuchsia [zircon-crypto] Start moving to //src/security/fcrypto
This will require a three-step transition to avoid breaking source
compat for an internal repo:

* Add `//src/security/fcrypto` target/code and deprecate
  `//zircon/system/ulib/zircon-crypto`
* Migrate internal, non-fuchsia.git usages
* Remove `//zircon/system/ulib/zircon-crypto` and transitional
  target/headers

This patch is step 1.

We shift the name from `zircon-crypto` to `fcrypto` because this
library, despite its name, is not used in the zircon kernel at all.
Its users are limited to zxcrypt (and other programs that link it)
and an internal tool.

We avoid the name `crypto` because that might conflict in library name
with boringssl's `libcrypto`, hence `fcrypto` for "Fuchsia crypto".

This commit moves tests, fuzzers, and all in-tree usages of
//zircon/system/ulib/zircon-crypto, but leaves a transitional group()
target at //zircon/system/ulib/zircon-crypto which redirects to the new
source_set along with including a transitional config for include_dirs
with include stubs to preserve #include compability for purposes of soft
transitions across repos as described above.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483441
Original-Revision: 6adaf82e6ab1ae07bda636ef2c04f0e9ff8c717d
GitOrigin-RevId: 0086d159cf46c54e1858257bfe72a92dccdc30b8
Change-Id: I1121fbda3fe256d3c4a4e89cc954098562e13717