[roll] Roll fuchsia [lowpan][lowpan-spinel-driver] Use empty identity when not provisioned
Previously, we were filling out the identity structure for the
lowpan network even if we were not provisioned. This change makes
sure that the identity is empty unless the interface is actually
provisioned with a network.

This change also updates `lowpanctl` to make sure that it prints out
the content of the identity structure even when not provisioned.

Original-Bug: 69572
Test: fx test lowpan-integration-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484458
Original-Revision: 521663f8ef4962bfc7803fdd2eb73fe5efc4c811
GitOrigin-RevId: 6b9957367ae463daa64310849e5404512795f623
Change-Id: I5f6633350557ac80163b922c38aff03317f670dc