[roll] Roll infra/recipes [build] Skip running buildstats if tool does not exist
We try to run buildstats even if ninja fails, but in that case ninja may
not have gotten as far as building the buildstats tool itself, leading
to confusing noise when buildstats fails (example:
https://ci.chromium.org/b/8855836404208460736). To avoid this, skip
running buildstats if the tool does not exist, but still add a log to
the "upload buildstats failure" explaining why we didn't run and upload
buildstats.

Also change the `api.step()` call to not use `ok_ret` so that failures
when running the tool itself will get caught and propagated to the
`upload_buildstats_failure` log.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/483377
Original-Revision: 7e6f64871d4ad74001636c1246ca79836a2a62c7
GitOrigin-RevId: ab1c0dbff71b1b100de419b01276bdd7302563f4
Change-Id: I48489d370723a937e7ca96d9c6225c1a2378804d