[roll] Roll fuchsia [system-update-committer] do commit when failure ignored
This change refactors the system-update-committer to make its behavior
consistent with the FIDL documentation for IsCurrentSystemCommitted.

Specifically, we actually complete the commit if a verification failure
is supposed to be ignored. This is different from the previous behavior,
which did not complete the commit upon an ignored failed check.

Test: system-update-committer-tests
Test: system-update-committer-integration-tests
Original-Fixed: 69537
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483138
Original-Revision: b0ecc2cc9a7ff3953ae842f3413bd15c310bf9e5
GitOrigin-RevId: 57bd213ad311ae45c4f2a4ceb28add98fcaccdcb
Change-Id: I399c1d3c485bb9c437a2bf9daca69f0c27dec6ad