[roll] Roll fuchsia [system-tests] Add pave test with old Zedboot
Test that paving using old Zedboot works with the current build.

This should prevent accidental changes that alter the image format.
Breaking changes to the image format are supported by submitting the
change to Zedboot to support it before changing how the image is
generated.

The test added here will be added to CQ in a follow up CL (which is in a
different repository).

Original-Bug: 66500
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480557
Original-Revision: aeea5c4dc281cf71f7514115b69d25fd1e2a97df
GitOrigin-RevId: cde369d7bfbe503a8ae27c5f59fb97b28608adfa
Change-Id: Ia33e9e826174a48d8d221fb0148705b84bd0edbe