[roll] Roll fuchsia [docs] Adds instructions about linking to in-tree code.
The style guide didn't have this piece of instruction, and it's
important because the static checks will fail without it.

Also, formatted the document to be 80 columns wide.  This is literally
what the document says the documentation style guide is, in the
section "Follow the 80 character limit".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480777
Original-Revision: 2564eaeb07698125426f3cb91c0118d80d6d4f43
GitOrigin-RevId: db75886747730320a9b48acd63faf7b529429400
Change-Id: I5984448872d607b83440ca481734149ff1e18b9c