[roll] Roll fuchsia [fidl][fidlconv] Support contextual built-in type resolution.
Previously, changes that required knowledge of the type underpinning a
particular identifier were impossible: if we're converting something
like "struct S { Foo a?; };", what does "Foo" mean?  We'd like to pursue
different conversions based on on the underlying type (ie, if "Foo" is a
struct, we'll need to box it, etc).

This change solves this problem by processing the file being converted
twice: once to generate a flat::Library, and a second time to generate
the raw AST for the ConvertingTreeVisitor to visit.  The flat::Library
representation is then passed to the ConvertingTreeVisitor, so that it
may use it to look up the underlying type when performing a
TypeConversion.

Original-Bug: 68667
Test: fidldev test fidlc
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481159
Original-Revision: c6e50082b571133bcb3e21f9c81235d7818bdfb7
GitOrigin-RevId: da84f3c9333bc5a03cb4f05b269c2cbbd8178a32
Change-Id: I73d0021b4d0a60260895b5ffb659b4b5b376026c