[roll] Roll fuchsia [fidlc] Emit obj_type for handles
Handles are mostly generalized, and introduced by `resource_declaration`
which specify an `obj_type`, itself an `enum` of a `uint32`. Exposing
the resolved value of an `obj_type` in the JSON IR to allow backends to
rely on this value, rather than duplicate the enum whose source of truth
is in a FIDL library (today, `library zx;` for Zircon handles).

In a next change, this new JSON IR value will be used to remove the
`ObjectTypeFromHandleSubtype` helper in fidlgen lib, along with
constants which currently duplicate the object type definition present
in the source FIDL library.

Original-Bug: 64629
Original-Bug: 45998
Test: fidlc tests and fidlc goldens
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477577
Original-Revision: 2a7216e086ce0daf4e147b5ecdfe75efec3dedad
GitOrigin-RevId: b7ae4dce1de1defab597398d785504604b5dbd0c
Change-Id: I32ee844e43832e2af8ce5e7ad8a34e1fe7a2545e