[roll] Roll fuchsia [paver]: Fix Unaligned Access.
This is cause for UB San failure. The problem is when accessing a
PartitionDescriptor/ExtentDescriptor, which is a serialized
struct(oacked) within a sequence of this packed struct, it is not
guaranteed that access is aligned.

Instead of reinterpreted pointer, memcpy to a aligned storage.

Original-Bug: 57974 #done
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484557
Original-Revision: b9285d0d74a63cb77b44500dc109438377819f76
GitOrigin-RevId: 46ca9dec5f3b1fe007ce9aad1a1e7ef4afc5816b
Change-Id: I1d8acca1927557b939cb00910d46f7487a0230a9