[roll] Roll fuchsia [wlan][policy] Recovers deleted interfaces.
This change checks notifications of deleted interfaces to
ensure that the IfaceManager was the one that requested the
deletion.  If the IfaceManager did not instigate the
deletion, the IfaceManager will attempt to recreate the
interface.

This is an initial attempt at recovering in instances where
the driver takes corrective action by deleting interfaces.

Original-Bug: 60089
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460756
Original-Revision: bb414fa37c3e339489a08d4cb7e1c0768b9d98e7
GitOrigin-RevId: 79988f1f0ed8f828f5705f17cec385ba373dbfa4
Change-Id: I003019b1b9a1a176118e91f1ea097bdc67374229