[roll] Roll fuchsia [scrutiny] Collector Scheduler - Dependency Ordering.
This change integrates a relatively complex scheduling ordering into
Scrutiny that forces execution ordering of dependent plugins collectors.

For example:
If plugin Foo depends on plugin Cores model during its collection the Core
collector must complete before the Foo collector can run. Since plugins
cannot be loaded unless all of their dependencies are already in the
tree this prevents cycles from occuring. Given this property it is
possible to resolve the tree one layer at a time as it forms a Directed
Acyclic Graph. Essentially all plugins with no dependencies are run
first, followed by all of those that have dependencies that exist in the
completed task set. This granularity applies at a plugin level only and
not at an individual collector level.

Internally state is tracked with a Mutex, Convar pair inside an Arc that
updates the state of the collector to the main scheduling thread.
Convars are utilized to allow the main scheduling thread to sleep as it
waits for the current batch of tasks to complete. This change also
introduces locking into the collector data structure itself so that when
the scheduler is running no collectors can be added or removed and more
importantly other schedulers cannot run. This simplifies the logic and
is fine for the use cases of Scrutiny.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483440
Original-Revision: efbd303d9ed8d6cb0763b0e3d8c46ce3dbe3f2ff
GitOrigin-RevId: ca94642a93e18f82028e99f717c4301fef5cc206
Change-Id: Ib3237eedc0bc10ce8f4dc3988aed000b56a566de