[roll] Roll fuchsia [fidlgen] Remove uses of ObjectTypeFromHandleSubtype
As we move towards properly generalizing handles, we can rely on the
object type emitted by the JSON IR in fidlgen backends. By doing so,
there are less ways for things to get out-of-sync. As an example, this
improvement fixed the handling of two object types clock and exceptions.
These are reflected in the goldens of the affected fidlgen backends (Go,
Dart).

While it would be nice to remove ObjectTypeFromHandleSubtype altogether,
we cannot do so quite yet. In GIDL, we parse handles with a nice looking
syntax like `#0 = event(rights: execute + write )`. There must be a
place to map `event` to its meaning, and since some GIDL backends --
looking at you fuzzer_corpus -- rely on the underlying object type
value, we must correspondibngly have a mapping to object types.

Lastly, since the mapping of clock and exception was missing, this was
added so that the converter is more correct.

Original-Bug: 64629
Original-Bug: 45998
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477578
Original-Revision: b536d66b3d6d83b93bd4550e9a406d0efd1e8f5f
GitOrigin-RevId: 373c8f275f53750044ec66c98e8211872354c0d5
Change-Id: I824767cfa2ab6160473fe7a3644decccadeac469