[roll] Roll fuchsia [fs] Derive PagedVfs from ManagedVfs.
ManagedVfs isused on Fuchsia for Blobfs to track the FIDL connections.
So if Blobfs is going to use PagedVfs eventually it needs to be derived
from ManagedVfs.

This design makes sense because ManagedVfs is designed to be run on
Fuchsia with FIDL, and paging will only work in this situation.
Long-term the paging functionality should be moved to ManagedVfs.

This also simplifies the test because ManagedVfs implements the
bookkeeping needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485582
Original-Revision: cdb69f9714a3423046465f5e88e2827057a1eadd
GitOrigin-RevId: 161494401235f2130bf515e97f22c1df8f14dfb9
Change-Id: I1d39bcbc561457680770b8b6fae29fde354ddac9