[roll] Roll fuchsia [sl4f] Check for null currentUrl from WebDriver
Sometimes a WebDriver session returns null as the current URL, which
throws an exception when passed to `Uri.parse`. I'm not certain of the
exact conditions required to get a null response, but it might be due to
the WebDriver session ending as the request for the URL is in flight.

Original-Bug: 68077, b/180029795
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485693
Original-Revision: da017a63919b3ca8448d33c6689c79f73949acda
GitOrigin-RevId: 0ff7c6cad8ed132ee89489cc2045c8f6dfe1ae1d
Change-Id: I7191ac69dbb91e45e65829442fe3b46a242ea9cf