[roll] Roll fuchsia [bt][common] AD::DecodeUri handles URIs that start with '\0'
URIs are length-encoded arrays of UTF-8 bytes (more like a std::string
made of UTF-8 bytes), not c-strings. As such, it is possible for a non-
empty URI to contain the byte '\0', which caused an assertion that
code_point >= 2 to fail. Remove this assertion (which is invalid);
instead handle the null byte like any other byte outside the range of
code points which correspond to encoding schemes. This allows for safe
removal of the special-cased uri.empty() check.

Original-Fixed: 69761
Test: `$ fx test bt-host-common-tests`, added a URI with explicit null
      bytes to GAP_AdvertisingDataTest.Uris.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485779
Original-Revision: 350b9c08cc9fa82333e3c56957af181484426ecc
GitOrigin-RevId: 9dc1df19eaa88a71f5e0d4805e20e94f8058f8ec
Change-Id: I8f23b922e087c4719926ee8ebbeef0783e6e6a5a