[roll] Roll fuchsia [fidl][cpp] Simplify variant handling
This contains a couple intertwined changes.

First, the template support functions now have the idea of a
current variant which is set in a template by calling {{ UseNatural }}
or {{ UseWire }}. Based on this there are now String(), Name() and
Namespace() methods on DeclType and a String() method on ConstantValue
and TypeName that return the appropriate value based on the currently
selected variant. It remains possible to explicitly request the Wire or
Natural variants for code that mixes them.

To support this the declaration types in fidlgen_cpp (Enum, Table, etc)
embed DeclName and no longer embed their raw IR (eg: fidlgen.Enum)
equivalents. That's because the raw IR contains a .Name field which
conflicts with the .Name method we want to offer to templates. Some
fields are lifted out of the raw IR types like Attributes and
Resourceness as needed. Type also embeds TypeName but that's a less
significant change.

So now to get the full name of a declaration in a template just ask for
its representation (eg: {{ . }}). To get its short local name ask for
.Name (eg: {{ .Name }}). The {{ EnsureNamespace }} helper understands
declarations now too.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484247
Original-Revision: 68156e04c6c0220006920c1b70cba4d5bc598b57
GitOrigin-RevId: 5aab74c25cf8b605c0bc10b0428e23f2aa23a4e4
Change-Id: I8a51aeb75978e1698731275bac173c0c3ad99a71