[roll] Roll fuchsia [codec_impl] Allow adapters to set profiles
The CodecImpl library uses a pair of threads internally: a shared_fidl
thread provided by the caller and an internal stream control thread.
Additionally, the DecryptorAdapter has an processing loop thread for
performing decryption operations. As a server library, CodecImpl is
largely unaware of performance or latency requirements, so we should
allow the Codec/DecryptorAdapter implementations set scheduler profiles.

Scheduler profiles can be acquired in a number of ways, but we should
allow CodecImpl users to use the least permissive approach, which is
fuchsia.media.ProfileProvider. That requires the thread handle to be
provided, so this CL exposes the stream control thread and the decryptor
processing thread to the Adapter implementations.

Original-Bug: 67050
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485520
Original-Revision: 794d562637b7be23760e8077d512bb9cad6161a7
GitOrigin-RevId: 31728a95ea1486f67f42688c092895c9f1bf1f54
Change-Id: I34b9116c97fcdd1cb64db2670cb60904d817b05c