[roll] Roll fuchsia [zxdb] Temporarily disable DwarfSymbolFactory.InlinedMemberFunction
Temporarily disable this test to facilitate the Clang roll. This test is
failing as of 1d68e0a0756476ceee1e62a778021e339c631940 because it checks
for the "this" parameter in a subroutine which Clang not omits because
it only has `abstract_origin` as an attribute. Clang is allowed to omit
this because it has no meaningful attributes.

After the roll, we will come back and re-enable this test with a fix.

Multiply: host_x64/zxdb_tests: 1
Original-Bug: 69724
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485156
Original-Revision: ba4dd52533d0444d39a6d941b121c5a7b7d87432
GitOrigin-RevId: 5ba85658d0916e40cbf12811b8b2cf5ba5b729e3
Change-Id: I9d2efda6881ba4bce10da23c89e08ac0dcad3f0a