[roll] Roll fuchsia [kazoo][go] Add fence between vdso{PC,SP}
In https://github.com/golang/go/issues/40823, it's hypothesized that
some test failures we observe are due to assumptions that, when `vdsoSP`
is set, `vdsoPC` is also valid. ARM may reorder stores (which is
possibly further exacerbated on multi-socket machines) to these fields;
insert a memory barrier between the production so that if `vdsoSP` is
observed to be set, `vdsoPC` is guaranteed to be observable.

This changes the syscall generation for Go, and relevant testing is
occurring there in the various multithreaded pprof tests.

Original-Bug: 69510

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485188
Original-Revision: 8538251176df69170ad2ddeb047a1490216d3ca8
GitOrigin-RevId: 818fb7edbb2d682a57582482f70b4305ad3e5c60
Change-Id: I325c9df472e47729212fc1ae71efc83792fc177d