[roll] Roll fuchsia [wlan][policy] Determine WPA3 support in policy
Adds a check to determine whether the device supports WPA3 in the
policy layer. This is used to reject connect calls to WPA3 networks when
WPA3 is not supported, and to treat mixed WPA2/WPA3 as WPA2 or WPA3 based
on whether WPA3 is supported. Network selection is not effected because
the lower layer already factors in  WPA3 support when determining whether
a network is compatible.

Multiply: wlancfg-tests
Test: added unit tests:
	fx test wlancfg-tests
      Manually tested that if WPA3 is not supported the connect
      call returns as not supported and if WPA3 is supported the
      connect succeeds. On reboot if a WPA3 network is saved and WPA3
      is supported we will attempt to auto connect to the WPA3 network.
Original-Bug: 67274
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460455
Original-Revision: bee8662af0aa60b64d0f961d0fe48796ec553cab
GitOrigin-RevId: 13d66bfdd789a467e003a2f0ff2d51ed6d53a850
Change-Id: Iee50ab1b5464c0815b2548d31b01debca058d789