[roll] Roll fuchsia [fint] Set --ninja-executable GN flag
For consistency with the current version of `fx set`. See
fxrev.dev/461355 for the full rationale of enabling it in `fx set`. It's
effectively a no-op on the infrastructure because infra always does a
clean build, and it doesn't significantly impact the runtime of `gn gen`
so there's no reason to not run it on infra as well as in fx.

Original-Bug: 68465
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484899
Original-Revision: fa611e0a4050f9f459c1d1bd6df35360c3f0a7d6
GitOrigin-RevId: d71519e9b9e252aadb290d35f1939ad63783da25
Change-Id: I545cc10a44700df9d8849a5af56248ae43ad3c45