[roll] Roll fuchsia [fs] Use STL locking, fix use-after-free.
The fs library previously used FBL locking because when it was written,
STL was not permitted at that layer. This prohibition has not been the
case for some time.

FBL locking can not be used in host builds and we have accumulated many
ifdefs around this. In addition, we plan to do more locking as we work
to make filesystems more parallel.

Standardizing all "fs" locking on the STL primitives means we can
compile on host or target and avoid problems with lock types not
matching.

The ifdefs around locks are kept to keep this patch only changes the
lock types, to reduce risk of landing. Some lock ifdefs will be removed
in a second pass.

The lock change revealed as use-after-free in
ManagedVfs::OnShutdownComplete where a lock was unlocked after the lock
ifself was freed. Fix the problem. This was revealed by asan which for
some reason was not triggering on the old types.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485744
Original-Revision: 70cf71b93555266dcdcc08bc7dfaad562f74156c
GitOrigin-RevId: 6aa386a54275879fca0deb7cfc2f04bc936fbb3e
Change-Id: Id10f1a37226528a63b01c52b501cebe0f750418f