[roll] Roll fuchsia [brcmfmac] Log channel info periodically.
This is a logging only change. As long as the client is
connected the current channel, rssi and snr are logged
every 5 minutes. This info is useful in the previous boot
log (where inspect data is not available).

Test: Since this is a logging only change, no new test
added.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485898
Original-Revision: 29b8d97e6408fd5def20386b1f7b8a6ad76dbe8a
GitOrigin-RevId: 943833a691e67ece7125ae8bcd630a5f4cfdd4c2
Change-Id: I45bbf30c662a7889d6bbed6c5c21590b733039d7