[roll] Roll fuchsia [bt][gap] Stop auto-connecting after persistent HCI errors
If we fail to connect to a peer (even after the retry logic) with any
of a set of specific HCI statuses (enumerated in ShouldAlwaysStopAuto-
Connecting), it's likely that there is a persistent issue preventing
the connection. If we have the auto-connect flag set for that peer, it
is possible for us to enter a loop where:
1. We see a connectable advertisement from that peer.
2. Auto-connect is set, so we try to connect to the peer
3. The connection to that peer fails, even though we are still in
   advertising range (e.g. due to asymmetric TX power).
4. GOTO 1.

To prevent this, add a step 3.5 (for auto-connect-initiated connection
attempts only), which sets the auto-connect behavior to not
auto-connect until after the next successful connection.

Original-Bug: b/178810924
Test: `$ fx test bt-host-gap-tests`, added the test
      GAP_LowEnergyConnectionManagerTest.
      ConnectSucceedsThenAutoConnectFailsDisablesAutoConnect, which
      emulates a successful connection+bond, & disconnect, then an
      auto-connect attempt which fails with one of the predefined
      HCI statuses and verifies that bt-host will not auto-connect.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485752
Original-Revision: 7f488701ca88a60ea3c775f66b9ddbd9289d1f9e
GitOrigin-RevId: 2acb59b554d20b4ca2fcde27feb99e2a85253aed
Change-Id: I7e4546fec9fa8a0a9c8e46f81b81dc2ee46acb6e