[roll] Roll infra/recipes [release] Use MILESTONE file to set major versions
Rather than hardcoding 0 as the major number, start respecting the
number recorded in the root of the integration checkout in the MILESTONE
file.

When new release versions are initialized, read the milestone number
from the checkout.

For example, if a new branch is being created from tip-of-tree main,
and the checked-in MILESTONE file says 2, then the new version spun up
on the branch will be 2.YYYYMMDD.0.1.

If an existing branch is being snapped from tip-of-tree main, it follows
the usual rules for incrementing the "release" (third) number e.g.
2.YYYYMMDD.R.1, where R is the existing R on the branch incremented by
1.

For snaps and cherry-picks that don't fall into either of these cases,
they will inherit whatever milestone number is currently there, e.g.
if an existing branch is on 1.YYYYMMDD.0.20, and is cherry-picked on,
this will create 1.YYYYMMDD.0.21.

If an existing branch is on 2.YYYYMMDD.0.40 and is snapped to another
branch, that branch will be on 2.YYYYMMDD.0.41.

Original-Bug: 70302
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/486267
Original-Revision: 312f22502b678a0e92a54cf8cce083e97a22c682
GitOrigin-RevId: 47a45f308c44b7953080127810cedab0e177364d
Change-Id: If558d260e6e672ee3a516e6da11a09fea775a794