[roll] Roll fuchsia [kernel][pc] Ensure EFI support exists on known-good platforms.
EFI is supported on most (e.g. NUC), but not all (e.g. QEMU), of our
x86_64 platforms. While we currently have tests to ensure that EFI calls
can be made, because EFI is not universal, the tests only do so if EFI
is detected.

One risk is that we inadvertently break EFI support in a way such that
EFI support can no longer be detected, but all our tests continue to
pass.

This test adds a list of "known good" platforms that are expected to
have working EFI support. If EFI cannot be detected on such a platform,
we fail the test.

Original-Bug: 69268
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483009
Original-Revision: be59c806fd5eba3579d5391b2868c70ee52d543f
GitOrigin-RevId: 0c99061d894831e171395433d4931784feef1c9a
Change-Id: Ice0e0a010693d3777a20ab642ad6c59f3d5d7cce