[roll] Roll fuchsia [kernel][lib/arch] Update LBR code
This change:
* Updates miscellany for style and clarity.
* Removes the dependency of lib/arch on fbl
* Rewrites all of the LBR tests. They were made overly convoluted in
aiming for maximal coverage of the iteration mechanics across the legacy
LBR formats, which ultimately has lead to poor RoI and readability. We
now exercise the logic that is directly a function of the legacy formats
in more direct unit tests and greatly simplify the scope of the iteration
tests. Further simplification is made in replacing the use of
hwreg::MockIo with arch::testing::FakeMsrIo.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477484
Original-Revision: 6e951092b211962e857ea3a8b6180a82398b3da1
GitOrigin-RevId: f3843bce8053e0f04318ffa0837311ab5ec235dd
Change-Id: I5829899e4dc0ccb76e16bc938bda3d6255c972ba