[roll] Roll third_party/go Regenerate Fuchsia syscalls
This change regenerates Fuchsia syscalls, which picks up a couple new
ones introduced since last time. It also updates the regen-syscalls to
work with current tooling interfaces.

The main point of this regeneration is to pick up the memory barrier
added between production to `vdsoPC` and `vdsoSP` in ARM64. In
https://github.com/golang/go/issues/40823, it's hypothesized that some
test failures we observe are due to assumptions that, when `vdsoSP` is
set, `vdsoPC` is also valid. ARM may reorder stores (which is possibly
further exacerbated on multi-socket machines) to these fields; insert a
memory barrier between the production so that if `vdsoSP` is observed to
be set, `vdsoPC` is guaranteed to have been written.

This change un-skips problematic tests in attempt to validate behavior.

MULTIPLY: go-pprof-test
Original-Fixed: 69510

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/485340
Original-Revision: b17886f06bc9e7686474277b49f786643141482b
GitOrigin-RevId: cde24c863360d01e6fb41a467e74a9248e97f7cf
Change-Id: If9414f7813b24ccd90b27f6b20b3185658af5042