[roll] Roll fuchsia [setui] Make DeviceStorage cache dynamic
This change removes the generic parameter from the DeviceStorage struct,
add generic parameters to the get and write methods, and stores an Any
in the cache in place of the concrete type.

This is in preparation for completely removing the single type storage
dependency on the controllers.

In order to minimize changes, PhantomData is being used in ClientProxy
types to limit the scope of type changes that are necessary. Those will
be updated in a future CL.

Original-Bug: 67371
Tests: fx test -o setui_service_tests
Also verified that storage still works by updating state and also
verifying it persisted across reboots.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485757
Original-Revision: 308b0c013287aa2ca2063605fab7e1712cdda89a
GitOrigin-RevId: 8ec8677bbdb8948a7defd6cfea084a02cebbda7c
Change-Id: Iccc06bfdd6bf5ae14e9f1082d587af5cb4e5f2b1