[roll] Roll cobalt [config_parser] Fix off-by-one error in calculation of num_index_points
In https://fxrev.dev/485978, the config parser set the num_index_points
field of a ReportDefinition to the width of the range of valid integer
values when the integer range is small. However, we want to include both
endpoints of the integer range as index points, so the value of
num_index_points was off by 1 in that CL. This CL fixes that error +
some golint errors.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/486957
Original-Revision: 3f48501cfbad60182a9d2afe630d41b96c504452
GitOrigin-RevId: ebcbfd63541d679c2dabb9855b20a9687c19e4a2
Change-Id: I30380b08e51177cb04aa262988a1bd05207347fb