[roll] Roll fuchsia [bt][gap] Always call pairing status callbacks
When connecting to a peer, it's possible for the pairing to complete
successfully, and then the peer disappears.  This manifested itself as
never getting called back BrEdrConnectionManager::OpenL2capConnection,
stalling connections from Profile clients.

Update the BrEdrConnectionManager::OpenL2capConnection API to accurately
reflect that the result always happens through the callback. Update
PairingState to finish incomplete local pairing requests when it is
destroyed, since it holds the callbacks while pairing is in progress.

Update the docs for PairingState to accurately reflect that the
status_cb on creation is not called on destruction, and the remaining
pending callbacks are.

Original-Fixed: 1414
Test: BrEdrConnectionManagerTest.OpenL2capPairingFinishesButDisconnects
Test: PairingStateTest.UnresolvedPairingCallbackIsCalledOnDestruction
Test: fx test bt-host-gap-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484860
Original-Revision: 62dcc5bd85a8f02a57a4c295a1de8e2cb16f0290
GitOrigin-RevId: 4940447ea847a5b3cfd8f24d26f5c9631213a9f4
Change-Id: I94db53b30993d6b5e25ec50440398eba654add15