[roll] Roll fuchsia [fs] Add locking to the PagedVnode.
Adds an inititial set of locking and lock annotations to the PagedVnode
to allow multithreaded access. The pager functions on the PagedVfs are
changed to allow the vmo_ access to be locked.

There is still a problem shutting down. The start of a new tracking
scheme is added with the addition of a Vnode::WillShutdown() function.
I will follow this up with some code in the VFS that tracks the
associated Vnodes better.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487078
Original-Revision: 8139fb7596ce4e8bee07203d896e8c391876cf90
GitOrigin-RevId: e8192d162d69d30319cf644ad83c00cb1b17a7d5
Change-Id: I9b713ae31ef3b3b2e35f9041b934400d16ee63a4