[roll] Roll fuchsia [bt][a2dp] Fix peers racing discovery to streaming
When we attempt to start streaming, the discovery and capability
collection process can take long enough that the peer starts
their own connection process after we have started collecting.

Previously we would attempt to start streaming and fail in this case,
after this change we will not attempt to start.

This is a reland of 480219.

Original-Bug: fxbug.dev/67408
Original-Bug: b/178721567
Test: added streaming_start_configure_while_discovery
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484861
Original-Revision: 847df5e78511d73d02709c25f09017651acacfe4
GitOrigin-RevId: 3a5f21529d5ec7565ed9e6372c1ee7e897dbc8fa
Change-Id: Id3a9935f5b74ef2b0cf1d840d9700b3c7a0badb3