[roll] Roll fuchsia [kernel][tracing] Add WaitQueue depth trace events.
Add a couple of trace probe events to WaitQueueCollection::Insert and
WaitQueueCollection::Remove which log the queue's pointer, and depth
after the operation.  These probes allow scripts to analyze traced to
determine what a typical N is for a wait queue, and how frequently
threads are added to or removed from a queue with a depth of N.

Because of the obvious potential performance impact, the tracing is
compile-time disabled by default, and can be enabled using the newly
introduced wait_queue_depth_tracing_enabled kernel build flag.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485593
Original-Revision: f540eb88ee9ab0b10060e3ff1962569c70fac0c2
GitOrigin-RevId: cfbbb1bcfffe8c56ce92c12272a8ab0f73a3a2ca
Change-Id: I3d7b4502bb6fe1439e228605ec3b44fa583ac3e0