[roll] Roll fuchsia [host-target-testing] Fix retry logic for artifact downloads.
The download() function would cleanup by removing the dst directory when
it encountered an error before retrying. This caused a problem when
using a srcs-file for blobs because the srcs-file was created as a temp
file in the dst directory so it got deleted before it could retry the
download. Another bug was that the blobs were being downloaded to the
`packages` destination directory which already contained some files, so
by deleting the entire directory on a retry, it removed previously
downloaded files in that directory. The fix downloads everything to
a temp directory first before moving it to the specified destination.

Original-Bug: 69467
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485218
Original-Revision: a7c7c6e7f253cdbda5bbbb394daae4d9839bc5b5
GitOrigin-RevId: 9697266914a1d24c88114c4acf75bc7450628f9b
Change-Id: Ic0863a8736ec07a19707a16279d0b1a076f5e06b