[roll] Roll fuchsia [blobfs] Fix BlobCompressor::Create and PrepareWrite to correctly return errors.
There was a bug in the current implementation of PrepareWrite,
wherein, even if BlobCompressor::Create fails, we always
return ZX_OK success. This is because, BlobCompressor::Create
returns a std::optional instead of also returning the error.
While this is good programming practice for C++, we ended
up introducing a bug as we always return ZX_OK even if
BlobCompressor::Create fails.
Ideally, BlobCompressor::Create should return the error
it encountered in addition to std::optional, but that is a much
more involved change, and I will do it as a followup CL.

Original-Bug:58097

Testing: Added BlobPrepareWriteFailure to test if we are returning error
         properly.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486557
Original-Revision: 99e0e05a417fd5b2c488f6d22c56b3721eca63ac
GitOrigin-RevId: 5233a0841d154af7430f02b2744e8ada4d825858
Change-Id: Ibbc09b27d9ff4e27e09938dd89e537b6888caf14