[roll] Roll fuchsia [wlan] Add WPA3 support to Authenticator
Now that Authenticator supports WPA3, a test for a bug in
Supplicant (fxbug.dev/69303) is added. The bug is fixed in this
change, bug full integration testing is postponed to
http://fxrev.dev/c/484547/ as it requires refactoring of hw-sim.

Added a _private field to fourway::Config to ensure a compiler error
will result if the fourway::Config::new() function is not used to
construct one.

Original-Bug: 69303
Test: fx test wlan-rsn-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485019
Original-Revision: 0fa99d46f75c600a8650fd488bc50a33b468bd4a
GitOrigin-RevId: 7dbaaf97c2c408b0b92614564c5f21baef1fe238
Change-Id: I68e9d000f31c85cffa5b8425de46ddb1aec7740a