[roll] Roll fuchsia [fx-set] Add support for build dir configuration
Decide which build dir to use depending on the value of the
_FX_BUILD_DIR variable, which is set if the top-level `fx --dir` flag is
set, and the --auto-dir flag.

I'm planning to stop supporting the --build-dir flag in this new version
of fx. It's been deprecated in favor of `fx --dir` for nearly two years,
with a warning printed on every use, so I think it's very likely safe to
remove. Supporting it would also make the already-complicated
conditional logic even uglier.

Original-Bug: 68465
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487275
Original-Revision: 71965a986b6a5f07b22ca831b4b7a78bb5243e16
GitOrigin-RevId: 7c33882902886d9759d016bb1a3d5faf49482872
Change-Id: Iccaf0fd955e2a2064b125d1f3a3c1da36b2c3c46