[roll] Roll fuchsia [a11y] Refactor GestureContext to include multiple pointers.
As we start to support more complex multi-finger gestures, we need to
pass multiple fingers' state to downstream users (e.g. magnifier). This
change modifies the GestureContext struct to maintain state for all
pointers in a gesture. Along the way, I also eliminated the GestureInfo
type, as it was redundant with the GestureContext and refactored some of
the state logic in existing recognizers to leverage the new
GestureContext.

BUG: fxbug.dev/69465
TEST: fx test a11y-lib-tests; fx test a11y-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482401
Original-Revision: d6ce0953cc34a11e169dbce1a71ff1ede9df6ada
GitOrigin-RevId: 61975e7879122e0bdcc6569e450bfca561c6042f
Change-Id: I0b6645e9bbef8b419c44697275ee0a716f84a9bf