[roll] Roll fuchsia [feedback_data] Don't FX_CHECK in reader
The reader was check-failing if it couldn't extract the number of times
a message was repeated from '!!! MESSAGE REPEATED'. Instead, if sscanf
fail, the reader appends the line it couldn't parse to the output and
continues on.

Original-Fixed: 70337

TESTED=`fx test system-log-recorder-tests`

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487244
Original-Revision: fbf69a761561b8df4cc1d9c1f9850991890599c5
GitOrigin-RevId: e4dbd043715a5de1fee385b83a0ef44cf27220a9
Change-Id: I5bf41eacbdbc310cdbac9051f79b17f979250f21