[roll] Roll fuchsia [gigaboot] improve disk_find_partition() API
disk_find_partition() was a bit weird in that it modified the passed-in
disk information to "shrink" the first and last block to just the
requested partition, so it no longer really represented the whole disk.
It also took in a partition name parameter, but only ever used it for
debug logging, so wasn't actually able to filter by name (or GUID).

This CL modifies the function to be a more general partition search
function, able to look for any combination of type/GUID/name. This will
be useful going forward to start adapting the logic to the new
partitioning scheme.

Original-Bug: 68692
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484382
Original-Revision: b942ac4be5611233e1593f00be2cab7584df560b
GitOrigin-RevId: 0e4d58a59efedf92ae669d73ed5939d80f88e299
Change-Id: I828307722618c52264b622067c0f5e411b911055