[roll] Roll cobalt Add flags to error_calculator_main
This CL adds four flags to the error_calculator_main binary and python
script.
  --simple: outputs a single value. Used to graph
            plots for go/tq-cobalt-error-colab.
  --max_value: overrides report's max_value field.
  --min_value: overrides report's min_value field.
  --max_count: overrides report's max_count field.

This CL also changes how an epsilon of 0 is handled by the calculator.
Previously, ε=0 was dropped as an invalid input. Now, ε=0 is assumed to
be no added privacy and always returns RMSE=0. Tests have been
updated to account for this change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/486262
Original-Revision: 6527aad12e962d24836ce55bdbcd0463012dce24
GitOrigin-RevId: 7702a2daa1e6e220ee1bccc9f4aaf213da3df3cd
Change-Id: I2c775f911a699a3d730aee1c10618ea433fe3df9