[roll] Roll fuchsia [vfs] Move lock to Vnode.
All minfs and blobfs files already had a lock so this will not increase
the object size. Directories and other uses of the VFS will now get a
lock that is currently unused, but these uses are rare.

This lock will be used to guard more data in a future pass (I want to
track Vfs pointers in the Vnode).

Change the Vnode::inflight_transactions_ from an atomic to a regular
value guarded by the lock. This should be almost the same and is
conceptually simpler.

There should be no behavior change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487400
Original-Revision: 15d7a14cfecc93f0bb82509eac9037d47b30be67
GitOrigin-RevId: 67a2c8c0cac212c5c96461cc40a75d2b129f252a
Change-Id: I925a7c8e2368e8fa7bf2a642f10aed89ace54b57