[roll] Roll fuchsia [connectivity] Migrate to new templates
The new templates do not support tags, so this removes a "flaky"
designation on wlanstack_bin_test and closes the tracking bug. It's not
clear if the underlying issue persists, but a new bug is probably
acceptable at this point.

Original-Fixed: 29873
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487917
Original-Revision: 8d2180b6352d988a411858c98a6673dde37dfa72
GitOrigin-RevId: 27e2264430e6e91589ffe6abfe8646383e5ab7df
Change-Id: Id84d46aada617006602b701d5711cf37af887219