[roll] Roll fuchsia [kernel][lib/arch] Introduce CacheConsistencyContext
Our arm64 version of arch::PostSelfModificationCacheSync was wrong (a
misunderstanding on the part of this author). This change replaces that
abstraction with an `arch::CacheConsistencyContext()` that separates out
the per-cache-line operations from per-cache operations in maintaining
cache coherence between the data and instruction caches.

Original-Bug: 67615
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477798
Original-Revision: f25366460d677582af26e065dd2fc622946294bc
GitOrigin-RevId: fc9c934393becb406c18235a4e4ab70f61109ae7
Change-Id: Idf6d95bd47c485b2d9ebd4e36980a79b219cdd4c