[roll] Roll fuchsia [vfs] Add Vfs backpointer.
Add a new constructor to Vnode that takes a Vfs pointer. The 0-argument
constructor will be deleted in a followup.

Tracks live Vnodes in the Vfs object and notifies them when the Vfs is
going away. This is like a weak pointer but we have access to the
backpointers.

Rename Vnode::WillShutdown to WillDeleteVfs. There is already a
"shutdown" operation for the ManagedVfs and that's a different step. We
don't want to delete the backpointers during ManagedVfs' shutdown. The
new name makes this distinction more clear.

Original-Bug: 70397

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487968
Original-Revision: ebd700691e96a1b4f775442139afc217c665dd69
GitOrigin-RevId: 0fb764c0d04e3aa84a07378524f56df1ad88d7fb
Change-Id: I1d260b5d4c8faf595be168b4a9ee3959bc69dde4