[roll] Roll infra/recipes [tricium] Always do a build when running against fuchsia.git
Previously, the tricium recipe assumed that it didn't need to do a full
build as long as all the files under analysis only required prebuilt
analyzers. However, this assumption is faulty because source files might
import files that are generated by the built (elements of the generated
source manifest) so we need to always do a build to ensure that the
generated sources are always available to the analyzer.

Unfortunately, this will slow down tricium by 5-10 minutes in cases
where we didn't previously do a full build, but there's not much we can
do about that.

Original-Bug: 70536
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/488082
Original-Revision: ea13a9f734332eaffcf7008e414e3e7a6607bf5e
GitOrigin-RevId: 3b50374697d8ad059abd5a1ed6921ec030a80532
Change-Id: I941044509a6f4e3a0831ef07c24ce8d13a4f24be