[roll] Roll fuchsia [stdcompat]: cpp20::span.
To simplify implmentation, we delegate the storage to an internal
struct, so we specialize that intead of the entire span implementation.

This CL just introduces the span construct, there are some constructs
missing such as:

cpp20::byte
cpp20::as_bytes
cpp20::as_writeable_bytes

Test are exhaustive and constexpr to verify that span works as intended
in constexpr context.

There are asserts for each undefined behvior, so bugs can be caught on
debug builds.

Introduced a new binary for testing, LIB_STDCOMPAT_NO_INLINE_VARIABLES
so this code is excersided in more environments, and we can more easily
recreate the conditions of the sdk.

Test: stdcompat-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483058
Original-Revision: 0d43fdc7e40129514bc353a30d32a2d9bf126e10
GitOrigin-RevId: f4cdc06bbde6cff814e2ca189a5641c921bd94b2
Change-Id: I9fc8ad14c77f21ee0924fa1faeb4dcbbe2f32270