[roll] Roll fuchsia [zxdb] Fix debug adapter server tests
accept() blocks indefinitely and does not exit if server socket is
closed after it is called. This leads to a race in ~DebugAdaperServer,
when it is trying to shutdown the thread. Using select() and pipe
to signal thread exit to fix this.

Original-Fixed: 66771
Original-Fixed: 66134
Test: fx test zxdb_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488209
Original-Revision: 168d4585373bec224968a6d73085d9d78a3934f8
GitOrigin-RevId: 500adfdbb4d0e3546698fc2b139d9e9a276f1748
Change-Id: Ica7bba675a49439d7145eb0147eacadefa692675