[roll] Roll fuchsia [events] Extract Events Client Library from OpaqueTest
OpaqueTest is on the path to deprecation. The events client
library is bundled with OpaqueTest but it doesn't need to be.

Extract this library as `component_events` so that OpaqueTest can
be safely deprecated.

`component_manager/testing:test_utils_lib` and `component_events`
are both marked as testonly. This impacts build rules that use
these libraries.

`step` is marked as testonly, otherwise it requires an exception
from these libraries.

Test: fx test hub-integration-test
      fx test storage-integration-test
      ... and more
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475897
Original-Revision: 2cd38659fb3f76fe26a8bf7183ecbb05a1120921
GitOrigin-RevId: 505b30c650e4a74f00af1c8e8e9826c831b4a23f
Change-Id: I8ecb9e0437fe67bfd0e30d0efef888ea0d8de625