[roll] Roll infra/recipes [release] Cleanup ref_to_release_version test data
Add a test_api for the "release" recipe module, which simplifies adding
test data for the many test cases that use this.

Also rename the "git describe" step to "ref to release version", and
change the presentation of the release version as a log to an output
property in checkout/api.py.

Original-Bug: 69550
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/488643
Original-Revision: 741a3c7c1824a680ed8a5ce1ec62fd14e434bf19
GitOrigin-RevId: 0f720af2a9b718963ede9777a43b23c376b07aee
Change-Id: I2f19235b8d607c30bfbad7294e2ace44acb9c4ef