[roll] Roll infra/recipes [recipes] Accept alternative names for commit-queue.cfg
CV, the replacement for CQ, will run tryjobs on behalf of a LUCI project
as long as the LUCI project has a commit-queue.cfg file. Previously
CQ would not run tryjobs unless the LUCI project also had a CQ instance
set up.

To make Sprout work, we use a single private CQ instance for all
repositories, but leave the public commit-queue.cfg in place so that it
can be fetched via the LUCI Config API by the recipes tryjob to figure
out which builders to run in recipes CQ. This will soon become
problematic because the mere existence of a public commit-queue.cfg
would cause CV to start running tryjobs based on that commit-queue.cfg
and race with tryjobs triggered off of the private commit-queue.cfg.

Therefore, we'll rename the commit-queue.cfg file used by recipes CQ to
not have the default name that CV checks for. To do that without
breaking the recipes builder, add a `cq_config_name` property to
override the default. Passing it as a property has the advantage that we
need only hardcode the file name in one place in the configs, rather
than duplicating the name between recipes and the configs.

Also configure the luci_config recipe to skip validating CQ configs for
projects that don't have a commit-queue.cfg file. This is easier than
plumbing the alternate file name into that recipe, and we'll still catch
all commit-queue.cfg errors because the public commit-queue.cfg is a
pure subset of the private one, which we'll still validate.

Original-Bug: 70617
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/488823
Original-Revision: 75f0564d0d952fb8cf1992b2aa8e9e67dd0bda12
GitOrigin-RevId: f06fa26fde7eb552a05dda4a65aa7897ff50bc0f
Change-Id: I4f176716470f311b7a3f0539261733eaaddbcd11