[roll] Roll fuchsia [codec] Free input packets shortly after stream failure
If a stream has seen a failure, including an async failure
from the CodecAdapter, any further input packets for that
stream are freed back to the client, and any packets already
with CodecAdapter (or racing) are freed either by the
CodecAdapter or by the subsequent
CodecAdapter::CoreCodecStopStream().

Test coverage for this is coming in fxrev.dev/486837.

BUG: fxbug.dev/70521
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488199
Original-Revision: cd08ffd26283ec0a9f65aaf21cd06c793d7c7e5f
GitOrigin-RevId: 8fb7ffc55002d33ff6e315f0552eb06cb73d7edc
Change-Id: I1087903eb0bbafc13b549009869d7c361c6d7f32