[roll] Roll fuchsia [codec_impl] Send OnStreamFailed before OnFreeInputPacket
It is largely expected for decrypting StreamProcessors to encounter a
stream failure on the first decrypt packet, as keys may not be available
yet. When this occurs, a OnStreamFailed event is sent and the stream
must be restarted. In http://fxrev.dev/486344, a change was made to
call OnFreeInputPacket for the packet that failed due to the lack of
keys to signal the client that the packet is free. However, since those
are independent events, there is no indication to the client that this
packet failed. In some h264 streams, if that first packet is not
retried, then the subsequent decoding attempt will fail as well.

For that reason, it is temporarily preferable to have the client assume
that any input packets and buffers currently with the StreamProcessor at
the point of stream failure were unable to be completed. The
OnFreeInputPacket events for those pending packets will be sent after
the OnStreamFailed event so that the client knows that they are
available for re-use.

Original-Bug: b/177173519, b/180649772
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486837
Original-Revision: 7ce815e76059ce8f85a0899b7e0a13763334e3cc
GitOrigin-RevId: 07a7b3d1305cef4747f9f23b904a10b8e1cbb29a
Change-Id: I48468579b410a5a5af6458e11976df97bc691c21