[roll] Roll fuchsia [debugger] Hook up thread registers.
The thread registers were not being passed through the unwinder to the
register listings. This meant that TLS values were not available. The
solution is to just copy the thread registers from the OS structure to
every stack frame under the assumption that they never change across
frames.

Adds a missing register record for gsbase on x64. This was causing an
assertion (in addition to meaning it wasn't available).

Original-Bug: 6570, 53940

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489028
Original-Revision: 0c1984c25a3f6e3b017c3eb0658fd4d61124af48
GitOrigin-RevId: c548746d82444b095a63573c7c477abd04bf70c4
Change-Id: I5e19d70626217737c142124e0e851d4fe68cf1b4