[roll] Roll fuchsia [devices] Ensure the root node is removed last in Node::Remove()
When removing a sub-tree, we may already be in the context of an unbind
operation. This means that we need to carefully handle the removal of
the root node in the sub-tree, so that all its children are removed
first.

Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488097
Original-Revision: a2da9f2cece49ee8df2d16a7c3e28e58a32370ff
GitOrigin-RevId: 21b68366a9e185b0243c749d3d5dd164f703bf25
Change-Id: I65a390ff3f96385f8875dcf33561206e89669313