[roll] Roll fuchsia [brcmfmac] Fix workqueue test flakiness
A workqueue test would sometimes access stack data that was no longer
valid because the test had finished but the access was running on the
default workqueue which had not yet finished. A previous fix was
attempted for this but it did not take into consideration that the
workqueue itself also accesses the WorkItem object after the work
callback has finished.

This fix uses the signaler mechanism of the WorkItem to ensure that the
test method doesn't exit until the work has actually completed.

Original-Bug: 70435
Test: Ran the test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487417
Original-Revision: 54f44618456c9056cf2899ad6a003c74be4b0e12
GitOrigin-RevId: 69356efa02174e0d95646257d1da147aac35e60a
Change-Id: I7dd2568d290d639fbe496267699c8df8fab2b7e5