[roll] Roll fuchsia [build] Allow the system image to depend on fuchsia_package() instances.
This CL modifies how the system image manifest is built, in order to
allow fuchsia_package() instances that are dependencies of
//build/input:system_image to contribute entries.

Before this CL, only the content of package() instance was recorded,
using the system_image_rsps metadata key, which point to a '.system.rsp'
file that lists the data in a format compatible with the final system
manifest file.

The CL introduces the system_image_fuchsia_packages_list() template,
which is responsible for collecting the fini manifests of all
fuchsia_package() dependencies, and converting it into a .system.rsp
format as well.

The general for testing is to do the following:

- Build, and save the content of
  out/default/obj/build/images/system_image.manifest somewhere,
  this file lists the content of the so-called "system package",
  i.e. all files that will be installed under /system/.. in the
  system image.

- Look at out/default/gen/build/images/system_image.manifest_args
  which will point to .system.rsp file for packages that contribute
  to the system image.

- Pick one such package, and modify the corresponding BUILD.gn file
  to replace package() and/or driver_package() with fuchsia_package()
  instead.

- Rebuild, and verify that the content of system_image.manifest hasn't
  changed compared to your backup.

Future CLs will just do this kind of changes for misc package()
instances.

Original-Bug: 70622

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488337
Original-Revision: 66165242f6ac311bf248a3dd718fe6046712b7c3
GitOrigin-RevId: bb1053d3fe9db939aa4ae73e76eb1a6f715eff0b
Change-Id: I7275bcbbf8db9ca76c6a38db32f43e63cecabdac