[roll] Roll fuchsia [rust] Set output_name for rustc_binary calls experiencing normalization
In preparation for a change to rustc_binary which will make output_name
default to target name without "-" -> "_" normalization, set output_name
on every binary currently experiencing that.

This change is a no-op; the generated binary names are staying the same;
the current binary files have _ in the name.

This CL is not meant to imply that the binaries _should_ have `_` or `-`
in the name, it's merely a reflection of the current reality. If you
want the binary to have "-" in the name, you can actually do that now by
setting output_name as of http://fxrev.dev/450757 , but be aware that
anywhere that references the name of the binary (.cmx and .cml files,
for example) will need to be updated to reflect the new name.

Original-Bug: 64100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486477
Original-Revision: 9613cd30cdf57922dd42afb24f518ae532cb4df9
GitOrigin-RevId: bf498f246eb73442f1e69fef15d9df8ddb9e5102
Change-Id: I03edb280a8f5436a2465e8aedf7b28fda2b36abc