[roll] Roll infra/recipes [go] Use a non-forked version of Go
infra/3pp/tools/go is a Chromium-maintained package that contains a
vanilla (i.e. un-forked) version of Go. The only reason we previously
used the Fuchsia fork of Go was to make it possible to build the
godepfile tool, which required a Go toolchain with Fuchsia support. But
that tool will no longer be used as of fxrev.dev/489497 so it can be
deleted and infra will no longer need to build it.

Also pin the CIPD package rather than using "latest" so that infra can't
randomly be broken by Go updates.

Original-Bug: 70189
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/489879
Original-Revision: 792e6f2ce413f0aa238984841bb10c86d4fea6f7
GitOrigin-RevId: e9efb924039ca7a4d9f422ac5ee981fc122f4046
Change-Id: Ic401328575737bee4d48d349661ffce31a3650a4