[roll] Roll fuchsia [devices] Wait on peer closed of the Driver protocol
Closing of the channel associated with the Driver protocol is used to
indicate to the Driver Runner that the driver has stopped executing
within a Driver Host. We should watch for that in the Driver Runner and
perform the appropriate clean up.

Original-Bug: 33183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489457
Original-Revision: fbd4b65473dc1b38ea9d434e8fd44f94709f02ec
GitOrigin-RevId: 0ffd81e4233c21005be65c50dc0fd585918f2782
Change-Id: I40092a1763f51ad2b67e65c3e20a89db52b01146