[roll] Roll fuchsia [fidl] Enable handle rights in fidl-format
The motivation for this is that handle rights are appearing in a
number of test files now and if you run "fx format-code" in a
CL that modifies one of these files, it will fail because the
handle files are considered invalid. Further, often prevent
formatting of other files.

It is relatively low risk to enable this flag in the formatter
as it generally won't hide errors unless users were trying to use
handle rights anyways.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485439
Original-Revision: ae0229f3cc597384cda444efe4d4c7ca42f36686
GitOrigin-RevId: 2a5a897dd4b091bc963e9a75d0c48cf72a86b569
Change-Id: Ic04d5f92e973d67677bf352ff54aace9325f66bd