[roll] Roll infra/recipes [checkout] Get release branch from gitiles_commit.ref if applicable
If we're using a release version, don't derive the release branches from
the commit. Instead, use gitiles_commit.ref, which will be of the form
refs/heads/releases/* if the build was triggered by a release branch.

This is more ergonomic for downstream users, since the potential (but
unlikely) plurality of release branches is a confusing edge-case to
handle.

Original-Bug: 69550
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/490357
Original-Revision: a9aa1ee08d77076aad6d36ae6795412fd5e76460
GitOrigin-RevId: 3bcd9bd577142bd46880b0b643660b67159fa6c1
Change-Id: I07d8a3125236b460bc0197943f89d8a70d357ce7