[roll] Roll fuchsia [time][fake-clock] Use fake time for a timekeeper integ test
Adds a simple integration test that verifies that timekeeper restarts
a time source if it has not produced any samples in an hour.
This test case does not need UTC support from fake-clock, but future
integration tests for timekeeper will require it.

As this is one of the first Rust components using fake-clock, this is a
test of both timekeeper and fake-clock.

Original-Bug: 61500

Multiply: timekeeper-integration

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484549
Original-Revision: 377c99586a6cf0075e0eae3e3d34992dd82ec641
GitOrigin-RevId: 561c913fbf91481b78244ff5237caa2eeafc7627
Change-Id: I507768ce46acd9cc41ca62a684dda17abfc722f6