[roll] Roll fuchsia [gigaboot] add support for new partition names
Adds support for `fastboot flash` to specify either new or old partition
names. Getting this in early allows it to soak for a bit so that devs
will flash it to their devices, at which point we could start
transitioning to the new patition scheme if we want.

Original-Bug: 68692
Test: unittests, `fx flash-remote`, `fastboot flash ziron_a <img>`
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490004
Original-Revision: 013fc83654e8e058f9e3e8689919ecb8617fdf63
GitOrigin-RevId: b3bc7a7bfc0640252b4bb81a3856fd3a1e3c4eee
Change-Id: Icf9381935f9db0fddddec76c678f059b8ee04b33