[roll] Roll fuchsia [fidl][docs] Move source compatibility test READMEs to //docs
All READMEs are moved into a "compatibility guide" section
which is a sibling to the "developer guides". This removes
the top-level separation between "guide"
type documents and "tutorial" type documents, but it
seems worth highlighting this section instead of nesting
it inside the developer guides.

This CL also makes accompanying changes to the gen tool:
* docs are now generated directly to //docs
* There's a separate regen_toc command to generate the toc for
  the "Compatibility guide" section. We don't do this automatically
  to avoid either issues with modifying files or regenerating
  the toc every time a single test changes. Since users will
  need to manually check the ABI/API compat guide for any
  manual changes they need to make as a separate step when
  adding/removing tests, running a separate regen_toc command
  shouldn't be too inconvenient.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487458
Original-Revision: 637cc46537efa90ed8886a6cfe1a098f61639846
GitOrigin-RevId: 51c41521a7301a49227a437a729e90544e347d8e
Change-Id: I748c87431def53d8f1a8880bf22dc5ad07c90791