[roll] Roll fuchsia [kernel][heap] Enable sized delete, assert static size fits in heap block
Enable generation of calls to sized-delete operators; when a compiler
knows that an object being deallocated has a particular size or alignment,
it can call a version of the delete operator with the size specified.

Some allocators can optimize for this path and avoid header lookups;
our kernel heap allocator cannot - its structure would make this
somewhat difficult. Instead, assert that the passed-in size is <= the
size in the malloc block's header. This would guard against certain
corruption of the allocator free list - if we free an object bigger
than its header, we may have underran our buffer.

This assert has very minimal perf costs on x86 - most perfcompare tests
are unaffected, the ones that are are affected marginally.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466819
Original-Revision: d49ef9856bcdfcf6a3bacd83b6455ee59972b21b
GitOrigin-RevId: 6a6c9a1551fea05b50f7d4f0d6f3c11664a86a72
Change-Id: I7aa92e7cbcc203c71f591a8bdbe5b1d978c42ec2