[roll] Roll fuchsia [settings] Remove deprecated watch2
Watch2 was used as a transitional api to move from returning errors on
Get() requests to closing the channel on failures and sending an
epitaph. The clients have been soft-transitioned to Watch2(), and now
need to be renamed back to Watch().

Original-Fixed: 55719
Test: fx test setui_service_tests, setui_client_interface_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405074
Original-Revision: ecacedd6fe5096c33cd6b87d2581d8f8aa802474
GitOrigin-RevId: 780c41ce533f0db44c915a428004d3f6157bbbfb
Change-Id: Idd8e3e1b99a61d69ef290de8fd7a78b8b8df2ff9