[roll] Roll fuchsia [e2e] Extend timeout of snapshot_test.
The SL4F facade this test exercises passes a 2 minute per-component
timeout to the ArchiveAccessor when taking the inspect snapshot.
Timeouts of individual components become more common under load which
affects Archivist, which means that the facade will often wait 2 minutes
before returning.

Unfortunately, the Dart test default sets a timeout here of 30 seconds,
so the test fails any time an indvidual component times out. This test
only cares about getting the inspect from Archivist itself, so most of
the time the test should be considered passing if there are some
missing components.

Increasing the Dart timeout to 3 minutes allows for the snapshot itself
to take 2+ minutes and for the test to still complete successfully. If
we find this test's slowness is a problem then we'll need to consider
ways to tighten the timeout the SL4F facade passes to ArchiveAccessor.

Original-Fixed: 70566

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491302
Original-Revision: 1728a5c167164deef92e884e1b9290c2b184e64e
GitOrigin-RevId: b834363619cfcc1dff1c49f1d2356649cc511041
Change-Id: Ic9d5e048a59bab1b9e5549a969211013d707817d