[roll] Roll fuchsia [bind] Define fuchsia.platform DID for XHCI
This adds the corresponding value to PDEV_DID_USB_XHCI_COMPOSITE.

Right now, we have two DIDs for the xHCI controllers: a DID_USB_XHCI and
DID_USB_XHCI_COMPOSITE.  Of the two, the composite variant is
overwhelmingly used.  We don't actually need to DIDs here, and to clean
this up I'm assuming the more commonly used value, and leaving the other
(which is only currently relevant to the vim2) to be cleaned up later.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489277
Original-Revision: 9ac34a485231d21ebe7ec2a86b200251960e7672
GitOrigin-RevId: 9d7a2a720c532a2fc8d3ecdd331e3e57d2558129
Change-Id: I9350bf75d54fe63d82074188649dd9207a490904