[roll] Roll fuchsia [async] Add method to delay a promise until and object is signaled
Like the time-based delay methods that were submitted before this, this
method on async::Executor produces a promise that will not complete
until a specified condition, in this case a zx_handle_t being signaled
with the signal provided when the user creates the promise.

Like the previous methods, the waiting starts when the method is called,
so if the user signals the handle before the promise is scheduled, it
will complete immediately when it is scheduled.

Original-Bug: 70802
Test: async-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491438
Original-Revision: e1e9a4bf1768c09c75d7992c6f29cff8df0e7709
GitOrigin-RevId: 651318839fd894a2ba4fdad42f8f6ed51f754ed0
Change-Id: Id2ad3a9e9767af30bed46e98c235bad697ef57c2