[roll] Roll fuchsia [fidl][sanitizers] Remove dependency of zircon/system/ulib on old C bindings
Remove the dependency of debugdata.cc on the deprecated C bindings for
fuchsia.io for calling fuchsia.io.Directory/Open. Unfortunately, because
llcpp depends (in some paths) on new/delete, and this code is used in
libc, this code cannot depend on llcpp as a replacement.

As a not great alternative, and because this is the only place that uses
this protocol that has this restriction, simply copy a few constants and
the structure definition from the C bindings inline to allow the fidl
protocol to be interacted with. A test to confirm that the local copy
does not drift from llcpp is added to z/s/ulib/c/test/debugdata.cc.

Original-Bug: 68648
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/488000
Original-Revision: 177290261c1dc18e25dbbd28f744e94de7e3558a
GitOrigin-RevId: b2e3f74aaff02b931b03abe72a16eabe6b7139f9
Change-Id: I0ee6994ca9049b68be9672ad5d89301719524ae4