[roll] Roll fuchsia [debugger] Enable timeout for symbol downloading
It's suspected that the hang of the symbolizer on infra bots is caused
by the network. Adding a timeout for symbol downloading makes it more
robust. The timeout is set to 10s for connecting and 10s for reading,
which should be sufficient for today's network.

Original-Bug: 66920
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491121
Original-Revision: 289031eb551816f1f63ef4c25ba51e3caad44414
GitOrigin-RevId: bc6a37ab320446afaceb4bb00497f24527cf108c
Change-Id: I39411b98857b686d493a88fdfc427f30aec50b87