[roll] Roll fuchsia [media/vnext] new fidl libraries for media vnext prototyping
This CL adds some fidl libraries to //src/media/vnext/fidl. These
definitions are a 'work in progress', but will be checked in so that
members of the media team can write code against them to refine and
expand on them. The intention is that they will eventually move into
//sdk/fidl and become part of the sdk.

fuchsia.media2 has the '2' suffix so that the 'one definition rule'
is not violated. When this becomes part of the sdk, we will need a
transition plan that allows us to replace the current fuchsia.media
definitions.

Definitions in fuchsia.media2 and fuchsia.mediastreams are based on
design documents currently under review. Definitions in fuchsia.audio
have not yet been subject to review and are speculative.

Test: no tests - fidl definitions only

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491578
Original-Revision: 5b14ca8e621d58dd7fbb3cec830a6335a7146fe4
GitOrigin-RevId: e08fdab684132581db24877d60f1f26b601914af
Change-Id: Ib2680556c1be7096281bcfeec1de2ee50ed9597e