[roll] Roll fuchsia [inspect][rust] Implement an API for performing atomic updates in the VMO
Introduces an API that increments/decrements the generation
counter a single time for a set of N operations instead of N times.
This gurantees that the group of operations will always be visible
together when reading.

Unlike the approach implemented in https://fxrev.dev/459554 this
approach doesn't guarantee thread isolation. So operations happening
in a separate thread while one thread is executing a set of
instructions will also be visible in the VMO.

Original-Fixed: 50060
Tested: added tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490011
Original-Revision: 8188f762fcd762ad5d2ebf52508cbb833525258d
GitOrigin-RevId: 0b15f65fe25f2e9686df5dd2245f3e8ce2c39239
Change-Id: Ifad2b443a462d21d833d50a19d55e4c5bfe9536c