[roll] Roll experiences [e2] Deflake ermine shell test take 2
The test to see if Recents is visible needs to take into account
that the terminal view is shifted to the right if Recents is visible.

The test was failing because Recents is a toggle, it can be cancelled
if focus is moved anywhere in the system (and other reasons). Checking
if the view rect is shifted to right, if Recents is visible, is more
robust.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/492002
Original-Revision: 50860203d289ad8802fc28575bbac4a57d4b9984
GitOrigin-RevId: 1a64ccd0f646d19e645a95e145849f62e1634320
Change-Id: I495bedff361a10b3beaa37c7c396326721977b9f