[roll] Roll fuchsia [rustc_binary] Set output_name on factory-reset-trigger
I missed a spot in http://fxrev.dev/486477 . Just like that CL, this
change is a no-op; the generated binary name stays the same; it's
currently named "factory_reset_trigger" because of the normalization
rules applied to `name`.

As far as I know, this is the last one.

Original-Bug: 64100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490198
Original-Revision: ba49afa5f46086e108194ba3ca40b087fe2499fb
GitOrigin-RevId: 6de69def5ab304721f3f231a8418de39eb95cad2
Change-Id: I484f4a014fa5c7c4bb9b8ac1b6351586e2d3d2fa