[roll] Roll fuchsia [flatland] add and use PresentArgs, squashable flag
This is the first in a series of CLs to convert Flatland to use Present3
semantics, as defined in http://fxbug.dev/69551.

This change has Flatland use a table, PresentArgs, for Present(), as
opposed to individual fields. This increases the flexibility of
Present(). No field is necessary, and if a field is omitted Flatland
will use a reasonable default value.

This change adds a field to PresentArgs called squashable. Presents with
|squashable| set to false can not be combined with subsequent Presents,
regardless of render timing and presentation times.

Original-Bug: 69551
Original-Fixed: 45945
Test: Unittests added. All unittests pass.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490721
Original-Revision: b5e517e483fe4f5947eeadb77111740fa9616564
GitOrigin-RevId: faea4c544fb2ac1287c256695a2272c00e64b4ff
Change-Id: I30d9ee5b1e4b99b43c9f1c11a18abbe7d1ca4fcf