[roll] Roll fuchsia [component_manager] Wait for shutdown to mark component deleted
This is a lifecycle behavior change that causes a component to be shut
down before it is marked deleted. ("Marked deleted" means the component
is hidden from the tree, but it may still have resources to be cleaned
up.)

The most visibile side effect is that Realm.DestroyChild will wait for
the child to stop before returning.  Note that with the new behavior, a
child cannot be recreated with the same name until shutdown completes.

This should also yield more intuitive system behavior. Users do not need
to be concerned about "zombie" components that continue to run and have
side effects but are not visible through normal means.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489417
Original-Revision: 61b65e8c2a6a424c8cd2bf2b05d4f7bf40b236ad
GitOrigin-RevId: 934ac6989416343b9a0e3bf28a0c75059d419177
Change-Id: I15f34f3a4b1dee68eff72aca28806e7ce33596b8