[roll] Roll fuchsia [brcmfmac] Fix re-connection issue after rebooting from WPA3 connection
This change fixed this issue by issuing an BRCMF_C_DISASSOC command
to firmware when an association is rejected or there is an
connection timeout. This is to clean up the connection info in AP of
our device.

Also this change clean up disconnecting status bit when the
BRCMF_C_DISASSOC is failed to send to the firmware.

Adding a delay in policy for WPA3 connection retries is also needed.
This is to bypass the WPA3 protection for re-association spam attack,
and is implemented in fxr/490941.

Original-Bug: 69067
Test:
  - Manually verified with donut, steps:
    1) Save and connect to WPA3 AP.
    2) Reboot the device with "dm reboot", verifies the connection
    status.
  - ACTS test case which covers the same steps.
  - Changed the unittest in sim/test/assoc_test.cc to verify the driver
  behaviour change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/482460
Original-Revision: 7d2d63c0f4153c443c1fe53a6a562e0acf9e0ca0
GitOrigin-RevId: dabcc37dd72e0d078d2af85f447b632b9e5e822b
Change-Id: I728f2ee0c55523990cbac5aadc42bf5c8caca214