[roll] Roll fuchsia [fidlc] Fix handle rights in coding tables
In the existing version, handle rights are not included in the
disambiguation calculation for handles in coding tables. This led to the
wrong handle type sometimes being pointed to in the coding table.

Additionally, disabled handle rights GIDL tests are enabled in this CL
for LLCPP.

Original-Fixed: 69798
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490400
Original-Revision: 34943d7bd24bea3770514d1290906a57d218d4fe
GitOrigin-RevId: 2a885d494c67635a52ef9bbcd1edc9691f42f8d5
Change-Id: Idc7acfa513f5237cf08f3c18e6ad6ba67234d7db