[roll] Roll fuchsia [tee][optee] Instrument yielding SMC calls
The OP-TEE driver makes a yielding SMC call whenever it is attempting to
communicate with a Trusted Application. Yielding SMC calls in the secure
world will operate with interrupts enabled and will return to the
non-secure world when one occurs. In order to detect whether this return
is taking too long, this CL adds a warning log whenever the peak SMC
call duration for a Trusted App call exceeds a given threshold. This
threshold is currently set to 10ms. Overall, the Trusted App call could
be much longer, but we are only trying to detect if a single SMC call
exceeds the duration.

Original-Bug: 69037
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489889
Original-Revision: 410c3aaba63301e8b80983a18cfa27f3c97225eb
GitOrigin-RevId: 88b00b5a07022460bcf8a3209ca1f1030326fc12
Change-Id: I6c2cd443b98e2110d7f4b581a00598e4a2fe06ca