[roll] Roll fuchsia [fidl][llcpp] UnownedClientEnd can return unowned_channel
Initially when defining UnownedClientEnd, we only used zx_handle_t
because zx::unowned_channel had a lot of baggage for what is a simple
handle number.

It turns out that there are lots of existing code that expect
zx::unowned_channel. It seems worthwhile to support returning a
zx::unowned_channel from an UnownedClientEnd. The getter that only
returns a zx_handle_t was renamed to `handle()`.

User code that obtains a handle just to call fdio_clone_service on it
are migrated to use `service::Clone`. In other cases, they are migrated
to use the `handle()` getter.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/489447
Original-Revision: b4d700fd8e430684c91e4a8108b3c630b2092bc7
GitOrigin-RevId: 02f4ceaba2f741419e9c5c367bae0f098b612d98
Change-Id: Ie7d1b3a39989e91c677903d609375a06b06f333d