[roll] Roll fuchsia [fx] add fastboot tool to bringup
Some flash workflows depend on fastboot, which is copied from prebuilt
by the //build/images:fastboot GN target. This CL adds this target to
bringup so that the tool is copied without requiring special '--with'
arguments, and added to the host tool metadata file (tool_paths.json).

Note that it is not enough to add it to the host labels, like
//tools/devshell:fx, because the build_api_module walks the
metadata only on the $default_toolchain, and thus the tool would not be
expressed in tool_paths.json as it should be.

Original-Fixed: 71094
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492641
Original-Revision: 9f7d94be58a894a77629aca4349ca3999bdecb7c
GitOrigin-RevId: 2298d57eda13d87751d94ae8b767d24df14791d1
Change-Id: If63329a205e55ba1cc0838bf10a80175641b4836