[roll] Roll fuchsia [time] Cease verifying the kernel UTC clock.
This is one step towards removing the zx_clock_get and zx_clock_adjust
syscalls as per RFC-0008.

Running kcounters on a test device shows we've successfully removed all
users of the kernel UTC clock, except for the calls to monitor that clock.

The CL removes that monitoring along with the deprecated calls we
needed in fuchsia-zircon to maintain it. This allows us to finally cease
setting the kernel UTC in the next step.

Original-Bug: 64063, 60521
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490584
Original-Revision: 8f6194451ba2cdb98a1e01537a30958e5ffad22e
GitOrigin-RevId: c6775b47a519bdeb0271d0a2d33b8bf2ac70aa9f
Change-Id: If5c3eec4e403f92958c29e0637c17164f926f2c6