[roll] Roll fuchsia [component_manager_for_tests] Log to syslog in test
Tests explicitly set up the kernel logger. Although this is appropriate
for prod, we should not be doing this in the test because we want log
output to be reported with the test results, like other tests.

Also, remove a stale `with_unit_tests = true` on component_manager_bin
(the tests are declared in separate components).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487267
Original-Revision: 9193ec5845e2a2fe44cdcce3099442006824ca62
GitOrigin-RevId: f7b6924fb202619ead36a333b41862851473e6cd
Change-Id: I085c47425eadfac79ae9d0f944cad27ab4b8a32e