[roll] Roll fuchsia [swd-e2e-tests] Make artifact download more efficient
This makes it more efficient to download artifacts in a few different
ways.

First, when working with old blobs that still use the old
build-archives.tgz and packages.tar.gz file. Instead of trying to
download a file that doesn't exist until we hit our retry timeout, we
track what files are actually part of a blob, and fall back right away
if it doesn't include the expanded artifacts.

Second, it deduplicates the artifacts we need to download so we only
fetch them once.

Original-Fixed: 70989

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491658
Original-Revision: e59ff28cac063b2476d7d9ff3e656c16b8161959
GitOrigin-RevId: b8960dc744a4d8e0611c111be530a75fa8055654
Change-Id: I6a7291ee156e8bb2a4881a9216684976160e27ca