[roll] Roll fuchsia [fs] Move ulib/fs to //src/lib
Moves //zircon/system/ulib/fs to //src/lib/storage/vfs/cpp to conform to
the new directory layout and be next to the Rust VFS implementation.

There are several hundred files referencing the old location so
forwarding BUILD files and headers are left in the old location. The
users of this library will be updated in a second pass.

Original-Bug: 70795
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490973
Original-Revision: 0641a319592da68b65668971c6b0f825ab545b39
GitOrigin-RevId: 25f6634085c63b8a0bdebbd80d7d4725392bfabc
Change-Id: I4f0f8ea97366a6b9218c211eceb9bfa41d2f00c3