[roll] Roll fuchsia [runner] Enable subprocess tests in CQ
subprocess.Runner is a heavily used utility by infra, so it's worthwhile
to run its tests in CQ. They run subprocesses so they're not very
hermetic, but they're very fast (~0.01 sec).

`go test -race` caught a race condition in the context cancelation test:
the test tried to read from the stdout and stderr buffers after
`runner.Run()` returned due to a context cancelation. But
`syscall.Kill()` doesn't wait for the signaled process to shut down, so
the process might still be running and writing to stdout/stderr even
after Run() returned. The easiest fix is to not collect stdout and
stderr in that test case, since the command isn't expected to produce
any output anyway and it including it in the error message harms
readability.

Also:
- Use `exec.Command()` rather than constructing an `exec.Cmd` struct
  directly. `exec.Command()` does $PATH resolution for relative paths,
  so we can avoid specifying absolute paths to `echo` and `sleep`.
- Check values of return errors in the tests to make sure Run() is
  exiting for the correct reason.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493439
Original-Revision: 40ceb9a5f4cf1a5acf2c3999afb078790866f872
GitOrigin-RevId: dcb4dfe10664af7c439300d33f27a9fdc3deb9e3
Change-Id: I6f281fccb3c32a5b6eb1f6f668f3bf22d04b602c