[roll] Roll fuchsia [netstack] Use assigned address in IP headers
The DHCP client retains its most recently acquired address so that it
can request the same address from the server when it is restarted. This
same address was mistakenly being used in outbound IP headers (and in
link address resolution); this produced incorrect behavior in
init-selecting where the client had not yet acquired an address but was
using it in its outbound messages.

Update TestRenewRebindBackoff to properly initialize the client in
states other than initSelecting to exercise ARP.

Original-Fixed: 69328
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493058
Original-Revision: 681f66dba5882c944518fdf0af1b870f2f380d4d
GitOrigin-RevId: 9242d14bb2b373bbdc8e93bee98ea158d3319af4
Change-Id: I8ca98296a9547ea4c3bd35ea2fbcb2ee00067d6b