[roll] Roll fuchsia [zedmon_test] Require -fake_zedmon_relpath flag
Otherwise, if the flag is unset the test will try to run an empty
command and hang until the test timeout is reached.

Long-term it might make sense to rewrite the fake_zedmon tool in Bash
instead of Go so that it can be run in the test without first doing a
build.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493441
Original-Revision: 7f54e428577cd72604188f7c73eebe4f08391454
GitOrigin-RevId: 7e7831e0b38fbb722e51c24609e0f2ededc972ad
Change-Id: Ibd164609dbb08198bcaef1eeeedff14bac2756bb