[roll] Roll fuchsia [gidl] Decode vector / array tests for handle rights
This CL checks the decode path for some known corner cases when handle
rights and types are checked in vectors. Specifically, in some bindings
there is no place to store rights and type information on the objects
themselves so they need to be passed down in another way.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493278
Original-Revision: 575de270208252e88c32b8dcd29881daf106afc0
GitOrigin-RevId: 85cf76150f7d8b8144868a43004195d891662a77
Change-Id: I2124183f69df02f68fd7aee7f48ac7420759b024