[roll] Roll fuchsia [kernel][x86] Refactor swapgs bug logic
We pull it into lib/arch to make it phys-friendly to enable its use in
code-patching in a follow-up change.

The previous swapgs logic determined what was affected in terms of
microarchitecture and had been written before the official CVE. Upon
rereading the official briefings, we now see that we actually expect
~every Intel CPU to be affected (with no indication that this will ever
be fixed) and so we can get away with a simpler "is Intel?" check.

Original-Bug: 67615
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487867
Original-Revision: 3ce4f5424cbfee2c91c84f8d662a7f35aa23a5db
GitOrigin-RevId: f7e3bdb5ff2dd7039ca8857edca24c2f433c52c9
Change-Id: I54edb564d08a753b76389b7756add7eb78485fc5