[roll] Roll fuchsia [cm][routing] Stateless routing
This change makes the routing code stateless. What this allows
is for routing to begin at any point in the process. Routing can
start at an offer, or an expose.

This opens the door for multi-step routing, where a capability is
routed to some intermediate declaration, where further action may
cause it route to completion. An example of this is Unified Service
routing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486383
Original-Revision: 959ee6ed8cb56551b266ed32dce19d0d3bf1415e
GitOrigin-RevId: 6fac1a103717826330ba6370e108ebbd1c542364
Change-Id: I9216510f23bf088c43f4775897f0850c1cb19fa7