[roll] Roll fuchsia [wlan][brcmfmac] Async remove ifaces as part of fw recovery.
This change causes the driver to remove the client and AP interface
(if it exists), as part of the firmware reset process.

Removal of the devices will ensure that policy layer is notified of
the devices going away (via the device watcher) and therefore it can
take corrective measures - i.e. recreate the necessary interfaces in
this case.

Original-Bug: 59364
Test: Test added. Triggered fw crash to ensure the expected sequence of events
played out.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/457533
Original-Revision: d95eb70827622aa5a970501367ab107babf56af6
GitOrigin-RevId: b6da00e4710b6839553e1d0d62b27c67f6af8b60
Change-Id: Ia4bbb2bde259e938524bac8a3be9397212cad5ae