[roll] Roll fuchsia [stdcompat] Add cpp20::invoke and related traits
This adds a conforming implementation of std::invoke as defined in
[func.invoke] and [func.require] ¶ 1 and 2, and the following traits
that also depend on INVOKE:

cpp17::invoke_result
cpp17::is_invocable
cpp17::is_invocable_r
cpp17::is_nothrow_invocable
cpp17::is_nothrow_invocable_r

And their related typedefs and variable templates.

This CL is structured so that the function template declarations
for the 7 points in [func.require] ¶ 1 are available to type_traits.h
so that the traits can be implemented without needing to include
functional.h.

We define invoke() in cpp20:: instead of cpp17:: (the standard when it
was introduced) because our version is constexpr, which is in accord
with the C++20 standard but not C++17. However, the internals are
identical to the C++17 version, so the internal namespace is still
cpp17.

Test: stdcompat-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474405
Original-Revision: f266336ae25b1b5d179041fc717b416654dafd3a
GitOrigin-RevId: 76f1f593feeba8d195a41dcd6933463a4ec11762
Change-Id: Ibf78714880b42ef29e5de2888683da4d99459fe1