[roll] Roll fuchsia [kernel][vm] Use gen count to avoid re-querying pages
Typically it is expected that a copy needs to be re-tried due to having
to fault in a mapping for a different vmo, and that this VMO is not
being actively modified in parallel. Using the generation count allows
detecting this common case and immediately re-trying to copy without
having to re-walk the VMO hierarchy and page list.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493219
Original-Revision: 6a0e1d20fdd811c8d9bd4190c46c6df67e88bb1e
GitOrigin-RevId: 61a0faaf44a1f0cc4165c3dc9d73f37a612b4c07
Change-Id: I5f2b625f9dd6b90fe7afa51868e7079754f8daea