[roll] Roll fuchsia [timekeeper] More resiliance on RTC time parsing.
From fxb/70974 it looks like the RTC sent an invalid time (e.g. month=0,
day=0). Which led to a panic in timekeeper. Thats undesirable since we
see an RTC as only a "nice to have". This handles the error gracefully
and includes the invalid time in the error message so we should be able
to isolate the problem if it happens again.

Original-Bug: 70974
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493660
Original-Revision: e9984038c4f54196a575f4635aae9ec2ea5f8e89
GitOrigin-RevId: 99ce48bd215e4048b9f7c1ae9d289d3f081ef699
Change-Id: Ieb4388989bb39f0bccf1a70fdd079495dd00e722