[roll] Roll fuchsia [kernel][crashlog] Print crashlog status to klog on boot.
Print a single line with info about crashlog recovery to the klog in
the default crashlog recovery handler.  This will be helpful in at
least understanding the uptime of devices which fail in CI/CQ with WDT
triggered reboots as the info should be present in the captured serial
logs.

Make sure that we only do this once.  The current crashlog API is one
of those which fetches the crashlog by calling recover with (0, NULL,
NULL) once to determine the size, and then possibly again to actually
fetch the log data. We avoid the double print using a simple atomic
bool.

Note that something different is going to need to be done for EFI
based crashlog recovery, however right now we are just trying to get
some data about ARM devices which are watchdogging in CI/CQ.

Original-Bug: 70724
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490589
Original-Revision: 7e3ccb00808624b44e6622f84aa9e3e657a8f829
GitOrigin-RevId: 25ef03b54121e13e8fe215a18309d331b0c9222b
Change-Id: I0f683de43fba0bdf46b724602844eb6b83c21936