[roll] Roll fuchsia [fidl][fuzzing] Seatbelt for cpp FIDL bindings fuzzer
The fuzzer was reading a uint64_t and using it to bound the size of a
vector of handles. Naturally, this allowed for instantiating more
handles than the allocator is willing to allocate. This change bounds
the vector at 2 * max number of handles in a FIDL message (to test an
oversized but reasonable collection of handles).

Original-Bug: 71159
Test: fx fuzz repro fidl-fuzzers/conformance-fuzzer [testcase-file-linked-in-bug]
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/494238
Original-Revision: 605ae475d888128a49d960a023a5c9ac6d6966e9
GitOrigin-RevId: 9df6ad6b03146190b0301265d814db952d78298a
Change-Id: I64e1ddb8ba67ca6131a0b2a6c0565d2622a45d9b