[roll] Roll fuchsia [time] Verify time source is kept alive if a dependency is unhealthy
Adds a test that verifies that an inactive time source is not restarted
if it has declared that some dependency is unhealthy.

This also refactors the PushSourcePuppet and allows it to track the
number of connections it has received. Tests that previously relied on
Cobalt events to verify the restart of a time source now inspect this
number instead.

Original-Bug: 61500

Multiply: timekeeper-integration

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493664
Original-Revision: b40128fb89315452d39d887c44f7bbcc57dec113
GitOrigin-RevId: 9d7718bf2ca605879aaa40df00361e2192d5e757
Change-Id: Ibf7a7c584963c9a96b7aec6dcb630e3a0d221afe