[roll] Roll fuchsia [kernel][boot-shim] Add "no-op" ZBI boot shim
This is a test case for the code that will be shared with the actual
legacy boot-loader boot shims.  With later build changes, it will
also directly act as the x86 ZBI shim for compatibility between the
legacy fixed-address ZBI protocol and the modern ZBI protocol.

Original-Bug: 32255, 68762
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484539
Original-Revision: 1aa7085813975c50155dfd390d634522b8c0b24c
GitOrigin-RevId: 59493d7352e729af6afbe066cb5851f86583d606
Change-Id: I7fbb83bf5ce2e7720c9cb873d20c0ba3af59aff6