[roll] Roll fuchsia [kernel][x86] Migrate swapgs bug mitigation to v2 patching
This is the first migration of the v1 code-patching over to the v2. As
mentioned in the bug, we will migrate over to v2 within the kernel ahead
of physboot's schedule. The pre-vm init hook that currently does the v1
code-patching now does both. We migrate the swapgs bug patching first as
it is one of the more simpler cases.

The v2 patching should replace the three one-byte `nop`s of v1 with a
single three-byte `nop`.

Successfully boots on both an Intel NUC and a AMD Zen.

Original-Bug: 67615
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/483180
Original-Revision: 8329be21476d4f630d564637ba35ddf9566d57ab
GitOrigin-RevId: 9b129da8af0e1a4fe2e3d31492ef59c0e7dfd0bc
Change-Id: Ie58ed12d5e241454ebb3f2cfef4a96bd1856f53d