[roll] Roll fuchsia [settings] flush input writes immediately
If the user hard reboots the device shortly after changing the sw mute,
the pending changes don't have a chance to flush and are lost. This
changes to flush immediately so that they are not lost in this case.

Follow-up bug for considering other settings: 71292

Original-Bug: b/181571540
Test: fx test setui_service_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493987
Original-Revision: fe3c0781b012bf0cbeaf88a89d55512d16653c56
GitOrigin-RevId: 5e9334de728a59d3525616df5a70aa757ba74536
Change-Id: I85b157767cdf98abd1873cea3f808d2569357cb8