[roll] Roll fuchsia [kernel][asan][arm64] Add shadow + checks + poison (2/N)
This CL enables kernel asan in arm64.

The big change that enables everything is support for tracking shadow
memory, and that is different to how it is done in kasan for x86_64,
but closer to how it is done in userspace asan.

When asan is enabled, memory accesses are checked against the asan
shadow, so everything mapped in the address space needs to have
something mapped in its corresponding shadow.

The asan shadow is now trakced by a vmar that covers it entirely. This
vmar starts fully unmapped, and portions of it are mapped as new kernel
memory gets committed.

The following parts are manually added, as they are calls to
ReserveSpace:

* Physmap arenas
* Kernel Sections
* Peripherals

For regular mappings, hooks were added in `ArmArchVmAspace::Map` and
`ArmArchVmAspace::MapContiguous` so that they will add shadow mappings
for new committed memory.

Given that it's possible for the same shadow page to be reused (this
could be solved by a quarantine, or hwasan in the future), the shadow
mapping function has some logic to avoid extra work in those cases.

Original-Bug: 30033 Kernelmode ASAN for Zircon
Pair-programmed-with: venkateshs@
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/481863
Original-Revision: 36e11d3e287c39d2aa874f0cb77a18d2fad9cd21
GitOrigin-RevId: b40dbcd16426daaad2fc054f65da8664d6ecec80
Change-Id: I5ca0d594e3d19ff9d64600debb36c20758e1c89c