[roll] Roll fuchsia [page-table] Return more detail in LookupPage.
In particular, also return the actual PTE and the level it is on.

LookupPage is currently only used in tests, but we need the additional
information in an upcoming test to ensure the correct size pages have
been used in certain operations.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/494604
Original-Revision: b3cf388565b411c71ce81f03f9b13ed5411d7dcd
GitOrigin-RevId: 49c539c3ccc8c9c1cf149d6f05288f2470ae6481
Change-Id: Ifd42f4fe955e50cc087036c66fb8f4d74caf147d