[roll] Roll fuchsia [kernel][lockup_detector] Fix a gap in CS coverage
Recent changes to the critical section lockup detector made it so that
threads exiting a critical section were not responsible for reporting
long running critical sections. Instead, heartbeat check-ins became
responsible for noticing that a thread had started, but not ended a
critical section for a long time (default, 3 sec). This is good
because it allows us to detect permanent critical section hangs before
the WDT has time to fire.

Unfortunately, it means that we lost the ability to notice long, but
not fatal amounts of time spent in critical sections (such as SMC
calls).

Restore some of this capability by introducing the per-cpu concept of
"worst case critical section time".  Threads exiting a CS now check to
see if the time spent in the CS was worse than their worst time, but
they do not make any effort to report it.  Instead, they just update a
per-cpu value, and let another checker report the value.

Checkers will only report a worst case value (as an OOPS) when it
1) Exceeds the minimum kcounter bucket threshold
2) Is worse than the last reported worst case
3) The rate limiter will permit it.

Additionally, the worst case for each core can be queried using
`k lockup status`

Original-Bug: 70724
Original-Bug: 69037
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/494721
Original-Revision: dc1547f61db1826ee797a5d57f408174ec50d25c
GitOrigin-RevId: f127ca114775fdf45e35176c390e90dfe2366e69
Change-Id: I358f79668419ab90005032786cd47d688c5ae5a4