[roll] Roll fuchsia [debugger] Fix the unwinder failure for abort()
libunwindstack expects a pc adjustment in order to unwind correctly. We
used to disable this feature because we want the return addresses
instead of call sites, which caused the failed unwinding for abort().

This patch enables the pc adjustment in libunwindstack but also restores
the pc after libunwindstack returns, so that the unwinding is improved
but the original behavior is preserved.

There's still an issue with the display of the frames when the pc is
outside of a function, which could happen when the function is calling a
noreturn function as its last instruction.

Original-Fixed: 45306, 58504
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493001
Original-Revision: 200e433ff37b09aec6bb76891d065da0885ea2d0
GitOrigin-RevId: 99d4dbe2ffbb86c8128500f1b57749ee9d1ffcec
Change-Id: I7fcf03255cd1e521ddc52cb2139cbb5d46ca4970