[roll] Roll fuchsia [kernel][lockup] Add new cmdline params and update docs.
Add two new kernel command line parameters which control the "fatal"
thresholds for the critical section and heartbeat checkers.

Soon, if these thresholds are enabled and exceeded, it will trigger
generation of a crashlog with a reboot reason of SOFTWARE_WATCHDOG,
and reboot the system.  Note that there is no actual change in
behavior, yet.  These parameters are being introduced a bit early in
order to get out of the way of the boot-options train, and to make it
easier to pick these changes into the F1 branch (which will not have
boot-options support)

Original-Bug: 70724
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495060
Original-Revision: f5db2b8801d6707ab0f24c28b829d8179d2e75c5
GitOrigin-RevId: 04ce735a79d0b734033e232a44b79ca43fd8949f
Change-Id: I8a4fc8bd67accbe0b0cd07f5b6fa87e8ab5a406a