[roll] Roll infra/recipes [build] Always build prebuilt binary manifests
Previously they would only be built by recipes that called
`api.build.with_options()`, and not on any builders that called
`api.build.ninja()` directly. However, it's bad for those two code paths
to be divergent because then it's not clear which one to use. The only
reason to call `gen()` and `ninja()` directly instead of calling
`with_options()` should be when the recipe needs to select some extra
ninja targets based on the GN artifacts.

So now we always build prebuilt binary manifests, which should have a
minimal effect even on the builders where they're not needed because the
manifests contain only prebuilt debug symbols, and those are only
included in the checkout on a small handful of builders.

Original-Bug: 67861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/494741
Original-Revision: 35e21cda6fb7da9d4722df4be5f320aae84a8ddc
GitOrigin-RevId: 0627efdc5699f6c88c1c68e9dc14354b5ec1cc94
Change-Id: I319a475792677c6d0874a2644b91ed6ce436e642