[roll] Roll third_party/go [fidl][go] Support rights and type checking in conformance tests
This CL is used in conjunction with:
https://fuchsia-review.googlesource.com/c/fuchsia/+/493599

This adds support for testing handle rights and type in
testutil_test.go and several fixes to the encode/decode
process that are needed for the GIDL generated
conformance tests to pass.

There are a few TODOs for cleaning things up which
involve a migration which I will do after this CL and
the fuchsia.git CL are submitted.

Original-Bug: 41920
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/493283
Original-Revision: 260c2b244fdd7ee3939f1a0b2d79db4f5a299e70
GitOrigin-RevId: 4ddb280e9316442b053929473e145c21a0f8b883
Change-Id: Iad0336b4df4af47566d062222b046a1f106f0139