[roll] Roll fuchsia [go][gidl][fidl] Handle rights GIDL tests for Go
This CL is used in conjunction with:
https://fuchsia-review.googlesource.com/c/third_party/go/+/493283

This enables support for checking handle rights and
types in the Go GIDL tests. More or less all that is
needed for this from the fuchsia.git side is
populating handle rights information on structs used
for checking.

Original-Bug: 41920
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493599
Original-Revision: cea69cb26e328f59dd7379ff66b8ee6a39cbf447
GitOrigin-RevId: 8b6339af7b507572bb38d497ad66a6e2e35527c0
Change-Id: I3757a3b80a500746ffaf9fc2970827636378d235