[roll] Roll fuchsia [docs][wdt] Clarify some confusing wording.
Clarify some of the wording in the WDT theory of operation doc.
Specifically "may not" (indicating that something is not permitted to
happen) was used when "might not" (meaning that it may or may not
happen) was the intent.

Also fix a mismatch where a second person pronoun (You) was paired
with a third person pronoun (They) later in the sentence.  Address
this by eliminating the You, and replacing it with a more appropriate
specific noun (Developers).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/494498
Original-Revision: 5a91921586e0f17c1b1eb1a902fe501f6ab77d99
GitOrigin-RevId: 4d3afe1138e6a4e32aecfda2d5fc72bc5eedb764
Change-Id: Ife15bfd66c779930c8ffc021af7d1cdc2c3ba56a