[roll] Roll fuchsia [fidlc] Pipe syntax flag through to Consume step
This CL sets the groundwork for switching behavior based on a
per file syntax. The current code uses a global experimental
flag in such a way that it can be mechanically replaced with
a per file version boolean when that is supported in the syntax.
The comment on ConsumeStep in flat_ast.h explains how this
value flows through the compiler.

To actually make use of this, support for protocols, services,
and resource declarations are added for in the new syntax. This
code path will eventually be tested once we update existing
tests to run on both the old and new syntax.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492398
Original-Revision: d571133b09d5cc9f0edf3dc6092812008185695d
GitOrigin-RevId: 54f5c07ebf009ec7e3205ecc4964c49f4697bb49
Change-Id: If9b524f49a1359338ff873ef3c7f35a3bdc88aa2