[roll] Roll fuchsia [appmgr] Set up and test directory walker for component storage.
This is the first change with associated tests to set up per-component
storage metrics export. The follow on changes will be to push them to
inspect, convert the component path names to useful monikers, and get
the data out to cobalt.

Currently this takes a list of directories to sum and periodically does
so via the main dispatcher.

Original-Bug: 69017

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/491339
Original-Revision: c37099f484771ed9367cf5f8b97284cd78b19f91
GitOrigin-RevId: 51eba12057b967822431e1738d90028fb15f17ec
Change-Id: I7718566e64e68fa0200706e093c99f72b91ff215