[roll] Roll fuchsia [hlcpp] Handle rights support for HLCPP conformance tests
This enables handle rights and object type checking in GIDL-generated
conformance tests for HLCPP.

There currently isn't a mechanism to check handle rights and type after
decoding and the handle id changes if there are too many handles
because the handle is replaced. Because of this, handle decode tests
are temporarily disabled with a TODO. I plan to revisit this and build
functionality for rights checking, but I hope to make incremental
progress by getting the rest of the conformance test handle rights
functionality working.

Original-Bug: 41920
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495178
Original-Revision: e1fbc6259c3219715130c5057e9934ce5470e88c
GitOrigin-RevId: 322eb910e9a8911d83d1752c8319d3afa4b1378a
Change-Id: If9cd1154fc3b72a1e01d761a634dadc91ace4c97