[roll] Roll fuchsia [kernel][wdt] Add a suppress feature to the HW WDT
Add the ability to suppress all watchdog petting to the WDT pdev
interface.  This feature will be used as part of the implementation of
higher level software watchdogs.  Specifically, we are about to get to
a point where each core in the system will be periodically checking in
on all of the other cores to make sure that they seem healthy.  During
these check-ups, the checking core will also pet the watchdog,
operating under the theory that as long as there is at least one core
still functioning, we do not want to allow the HW WDT to fire.

If a checking core decides that some other core has become fatally
wedged, it will attempt to write a crashlog and reboot the system.
What happens, however, if this core wedges while attempting to write
the crashlog?  If there is a 3rd core which is still somehow petting
the dog, we could end up in a situation where we have lost multiple
cores, but he dog keeps getting pet.

In order to avoid this worst case scenario, we introduce a "suppress
petting" flag to the watchdog.  When a core detects another fatally
locked up core, it will:

1) Pet the dog one last time.
2) Suppress petting (now other cores attempts to pet will be ignored)
3) Create the crashlog
4) Reboot.

This will give the system one last HW WDT period (5 seconds on Astro
and Sherlock) to create its crashlog and reboot.  If this fails to get
done in a timely fashion for any reason, the HW WDT will still kick in
and reboot the system.

Original-Bug: 70724

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495987
Original-Revision: 924384ce7c5fa78733be90e5e952bc472222539d
GitOrigin-RevId: 13d3012006a08a17adb610c1fcfdfd55c3666003
Change-Id: Ideb8df135e551a5d3318f557bb7e1e32be97b17e