[roll] Roll fuchsia [fdio] Simplify ppoll and select
Use deferred cleanup to avoid various gymnastics.

This also fixes the behavior of both functions when zero fds are passed
and the timeout is a null pointer; this change is untested because the
correct behavior in that case is to block indefinitely.

This also correctly returns an error from ppoll when a negative timeout
is passed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495239
Original-Revision: 17fd63c51e2a5b78178b81455faedfac5eb8e352
GitOrigin-RevId: d534884ffbf66d8901609aa1b28fb3e04ef52efb
Change-Id: Iea8d7b977bab4c091e47f268b8548ba9f143084c