[roll] Roll fuchsia [wlan] Add MLME event that triggered disconnect to Inspect tree
The wlanstack.cmx Inspect tree no longer includes the
disconnect_reason field and has an expanded disconnect_source field.

The new format of the disconnect_source field is following:

  "source: <user|ap|mlme>, reason: <disconnect reason>"

If the source if ap or mlme, then the string will have the following
appended:

  ", mlme_event_name: <DisassociateIndication|DeauthenticateIndication>"

Including the mlme event that triggered the disconnect adds an extra
dimension which can be used when determine the type of event which
triggered a disconnect.

Test: fx test wlanstack-tests wlan-sme-tests
Original-Bug: 70687
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/496022
Original-Revision: 9d42ee332c82657d30bc5c578188262a611f01e2
GitOrigin-RevId: dabe6c39b55f84be752bf0391381445d86c08239
Change-Id: Ia64574c0a6b5d909695a19b5889fa5def18843eb