[roll] Roll fuchsia [docs][components] Non-base component resolver section
Add a section to the migration guide which describes what to do
when the component is not part of base package set. If a component
is in the cache or universe sets the universe resolver is required
to load the component.

Original-Fixed: 70891
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/492721
Original-Revision: f85414fb9be728030b8f669f1fb9f705aaa9855c
GitOrigin-RevId: ee23a8e362dfeda6c4d8fe3f4804199f922a82d8
Change-Id: I7a6732ee894f40adf02740e51ca44a3e292ca065